% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\docType{data}
\name{alerts}
\alias{alerts}
\title{CCAPS Alerts data}
\format{
A data frame with columns:
\describe{
\item{subscale}{CCAPS subscale}
\item{bin}{Bin for the baseline CCAPS}
\item{session}{Session number}
\item{score}{Score above which a CCAPS alert is produced}
}
}
\usage{
alerts
}
\description{
Provides a lookup table for each subscale, starting CCAPS bin, and session indicating the score above which a CCAPS alert is produced.
}
\keyword{datasets}
