% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.r
\name{apa_table}
\alias{apa_table}
\title{Create APA formatted tables}
\usage{
apa_table(
  data,
  header.names = NULL,
  title = "Table #",
  subtitle = "Add subtitle",
  save = FALSE,
  path = "table.png",
  width = NULL,
  note = NULL
)
}
\arguments{
\item{data}{A data file.}

\item{header.names}{A vector of header names to change columns names. Default is \code{NULL}.}

\item{title}{The title of the table. Default is \code{Table #}.}

\item{subtitle}{The subtitle of the table. Default is \verb{Add subtitle}.}

\item{save}{A logical statement indicates whether the table should be saved as a file under a local folder. If false, the table will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway and name if \code{save = TRUE}. By default, \code{"table.png"}.}

\item{width}{A numeric value between 0 to 1 to indicate the width of the table. By default, \code{NULL}.}

\item{note}{A quoted string to add a note at the bottom of the table. By default, \code{NULL}.}
}
\value{
Print and saves APA formatted tables.
}
\description{
Create create APA formatted tables.
}
