% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_CLI.R
\name{bin_capacity}
\alias{bin_capacity}
\title{Create a vector that bins capacity.}
\usage{
bin_capacity(x)
}
\arguments{
\item{x}{A variable within a data frame that pertains to capacity at the center.}
}
\value{
A vector with capacity bins.
}
\description{
Capacity is the total number of contracted/expected clinical hours for a typical/busy week when the center is fully staffed. This function is often used in relation to data frames that contain variables related to the Clinical Load Index (CLI). Capacity bins include "48 or less (0-2 Standardized Counselors)", "49-72 (2-3 Standardized Counselors)", "73-96 (3-4 Standardized Counselors)", "97-120 (4-5 Standardized Counselors)", "121-144 (5-6 Standardized Counselors)", "145-168 (6-7 Standardized Counselors)", "169-192 (7-8 Standardized Counselors)", "193-240 (7-9 Standardized Counselors)", "241-312 (9-13 Standardized Counselors)", "313-432 (13-18 Standardized Counselors)", and "over 433 (18+ Standardized Counselors)".
}
