% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_CLI.R
\name{bin_enrollment}
\alias{bin_enrollment}
\alias{bin_utilization}
\alias{bin_capacity}
\alias{bin_inst_utilization}
\title{Bin CLI variables (enrollment, utilization, clinical capacity)}
\usage{
bin_enrollment(x)

bin_utilization(x)

bin_capacity(x)

bin_inst_utilization(x)
}
\arguments{
\item{x}{Variable to be binned}
}
\value{
Binned variable
}
\description{
Create categorical versions of CLI variables (enrollment, utilization, clinical capacity)
}
