% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_CLI.R
\name{bin_inst_utilization}
\alias{bin_inst_utilization}
\title{Create a variable that bins the percentage of students who utilize a center at the institution.}
\usage{
bin_inst_utilization(x)
}
\arguments{
\item{x}{A variable within a data frame pertains to the percentage of students who utilize a center at the institution.}
}
\value{
A vector with institution utilization bins.
}
\description{
Institutional utilization refers to the percentage of students who utilize a center at the institution. This function is often used in relation to data frames that contain variables related to the Clinical Load Index (CLI). Institution utilization bins include "less than 5\%", "5-7\%", "7-10\%", "10-12\%", "12-15\%", "15-20\%", "20-30\%", and "more than 30\%".
}
