% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_CLI.R
\name{bin_utilization}
\alias{bin_utilization}
\title{Create a vector that bins utilization.}
\usage{
bin_utilization(x)
}
\arguments{
\item{x}{A variable within a data frame that pertains to utilization at the center.}
}
\value{
A vector with utilization bins.
}
\description{
Utilization is the number of students seen by a center. This function is often used in relation to data frames that contain variables related to the Clinical Load Index (CLI). Utilization bins include "under 151", "151-200", "201-300", "301-350", "351-400", "401-500", "501-600", "601-700", "701-850", "851-1000", "1001-1200", "1201-1500", "1501-2000", "2001-3000", and "3001+".
}
