% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_closure_key.R
\docType{data}
\name{case_closure_key}
\alias{case_closure_key}
\title{Case Closure Form key}
\format{
A data frame with columns:
\describe{
\item{Item}{Case Closure item corresponding to the column name of each item in the data}
\item{Name}{Name of the Case Closure item}
\item{Number}{Case Closure number corresponding to the top reason for ending treatment item}
}
}
\usage{
case_closure_key
}
\description{
Contains a key for the Case Closure form items and numbers.
}
\keyword{datasets}
