% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAPS62_cuts.R
\name{ccaps62_cuts}
\alias{ccaps62_cuts}
\title{Create CCAPS-62 cut points}
\usage{
ccaps62_cuts(data)
}
\arguments{
\item{data}{A data file containing CCAPS-62 subscales}
}
\value{
A data frame with all the original data in \code{data}, and several additional columns of cut scores coded 1 for above cut score and 0 for below cut score.
}
\description{
Creates binary variables for CCAPS-62 subscales indicating whether the score is above or below the low and high cut respectively, using the 2018 cut points.
}
\note{
If \code{data} does not contain the proper CCAPS subscale names, it will return an error.
}
\examples{
\dontrun{
TI1718 <- ccaps62_lowcut(TI1718)
}
}
