% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAPS_change.R
\name{CCAPS_change}
\alias{CCAPS_change}
\title{Upload CCAPS data with change scores between first and last completion}
\usage{
CCAPS_change(
  data,
  client_identifier = UniqueClientID,
  center_identifier = CcmhID,
  add_items = NA,
  include_first = F,
  include_last = F
)
}
\arguments{
\item{data}{A data frame with CCAPS data.}

\item{client_identifier}{The column uniquely identifying each client. By default, \code{UniqueClientID}.}

\item{center_identifier}{The column uniquely identifying each center. By default \code{CcmhID}.}

\item{add_items}{Specify the columns or CCAPS variables outside of CCAPS subscales to be included in the returned data frame. By default \code{NA} resulting in no additional CCAPS variables being added to the returned data frame. Multiple features should be listed in the following manner: c("Variable1", "Variable2"). Option \code{"all"} selects all CCAPS 34 variables to be in the returned data frame.}

\item{include_first}{A logical argument that specifies if subscale or variable scores from the first completion of the CCAPs should be included in the returned data frame. By default \code{FALSE} resulting in CCAPS subscale or specified variable scores of the first completion not being included in the returned data frame. Option \code{"TRUE"} adds first completion scores of CCAPS subscales and specified variables to the returned data frame.}

\item{include_last}{A logical argument that specifies if subscale or variable scores from the last completion of the CCAPs should be included in the returned data frame. By default \code{FALSE} resulting in CCAPS subscale or specified variable scores of the last completion not being included in the returned data frame. Option \code{"TRUE"} adds last completion scores of CCAPS subscales and specified variables to the returned data frame.}
}
\value{
A data frame with CCAPS change scores of subscales and specified variables. Based on additional arguments, scores from first and last completion could be added to the data frame.
}
\description{
Upload CCAPS data with change scores between first and last completion
}
