% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccaps_cut_labels.R
\name{ccaps_cut_labels}
\alias{ccaps_cut_labels}
\title{Creates label variables for CCAPS-62 and/or CCAPS-34 subscales cut scores.}
\usage{
ccaps_cut_labels(data)
}
\arguments{
\item{data}{A data file containing CCAPS-62 and/or CCAPS-34 subscales cut scores.}
}
\value{
A data frame with all the original data and additional columns of cut score labels.
}
\description{
The functions create label variables for each CCAPS-62 and/or CCAPS-34 subscale cut score.
}
\note{
To run this function, cut scores must have been created using ccaps34_cuts and/or ccaps62_cuts functions. Three types of label variables are created. Variables ending with "_sum" (e.g., Depression34_cut_sum) represent the sum of the low and high cut variables (e.g., Depression34_low_cut + Depression34_hi_cut), with possible outputs of 0, 1, or 2. Variables that end with "_sum_label1" (e.g., Depression34_cut_sum_label1) are labels based on the "_sum" variables where the data was recoded: 0 = "Low", 1 = "Moderate", and 2 = "Elevated". Variables that end with "_sum_label2" (e.g., Depression34_cut_sum_label2) are labels based on the "_sum" variables where the data was recoded: 0 = "Not elevated", 1 = "Not elevated", and 2 = "Elevated".
}
