% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAPSvalue_check.R
\name{CCAPSvalue_check}
\alias{CCAPSvalue_check}
\title{A Check for the Values of CCAPS Items}
\usage{
CCAPSvalue_check(dat)
}
\arguments{
\item{dat}{The data frame containing the new year of (combined) data.}
}
\value{
A list with two elements. The first, \code{violations}, is a boolean indicating
whether or not any violations were found. The second, \code{CCAPS_viol}, is a vector of
strings containing the names of the CCAPS variables that have at least one value out of
bounds.
}
\description{
A function used to check the validity of the CCAPS items. All CCAPS
items should have values between 0 and 4, and possibly NA for missing values. This
function checks whether these constraints are satisfied, and if not, returns a list
containing the CCAPS item(s) which violated these rules, and their associated
offending value(s).
}
\note{
\code{CCAPSvalue_check} calls on CCAPS variables by name, so they must be named
properly in \code{dat}. Variable naming convention is \code{CCAPS_01}, \code{CCAPS_02},
etc. If \code{dat} does not contain the proper CCAPS variable names (or the naming
convention was changed in the new year's data), the function will return an error.
}
