% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{ccmh_caption}
\alias{ccmh_caption}
\title{Format a ggplot2 object to add CCMH caption.}
\usage{
ccmh_caption()
}
\value{
A ggplot2 object.
}
\description{
This function will add CCMH's caption to a ggplot2 object.
}
