% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{CCMH_palettes}
\alias{CCMH_palettes}
\alias{CCMH_institutional_palettes}
\alias{CCMH_categorical_palettes}
\alias{CCMH_gradient_palettes}
\alias{CCMH_diverging_palettes}
\title{CCMH palettes}
\format{
\subsection{\code{CCMH_palettes}}{

An object of class \code{palettes_palette} with 28
colour palettes. Use \code{names(CCMH_palettes)} to return all palette names.
}

\subsection{\code{CCMH_institutional_palettes}}{

An object of class \code{palettes_palette} with
4 discrete colour palettes.
Use \code{names(CCMH_institutional_palettes)} to return all palette names.
}

\subsection{\code{CCMH_categorical_palettes}}{

An object of class \code{palettes_palette} with
12 sequential colour palettes.
Use \code{names(CCMH_categorical_palettes)} to return all palette names.
}

\subsection{\code{CCMH_gradient_palettes}}{

An object of class \code{palettes_palette} with
9 sequential colour palettes.
Use \code{names(CCMH_gradient_palettes)} to return all palette names.
}

\subsection{\code{CCMH_diverging_palettes}}{

An object of class \code{palettes_palette} with
3 sequential colour palettes.
Use \code{names(CCMH_diverging_palettes)} to return all palette names.
}
}
\usage{
CCMH_palettes

CCMH_institutional_palettes

CCMH_categorical_palettes

CCMH_gradient_palettes

CCMH_diverging_palettes
}
\description{
Palettes for CCMH.
}
\examples{
# Get all palettes by name.
names(sunset_palettes)

# Plot all palettes.
plot(CCMH_palettes)
}
\keyword{datasets}
