% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{ccmh_theme}
\alias{ccmh_theme}
\title{Format a ggplot2 object with the CCMH theme.}
\usage{
ccmh_theme(
  plot_title = 20,
  axis_title = 16,
  axis_text = 14,
  caption_text = 12,
  strip_text = 14,
  text_font = "Avenir"
)
}
\arguments{
\item{plot_title}{A numeric value to indicate the plot title text size. By default, \code{20}.}

\item{axis_title}{A numeric value to indicate the axis title text size. By default, \code{16}.}

\item{axis_text}{A numeric value to indicate the axis labels text size. By default, \code{14}.}

\item{caption_text}{A numeric value to indicate the caption text size. By default, \code{12}.}

\item{strip_text}{A numeric value to indicate the facet strip text size. By default, \code{14}.}

\item{text_font}{A quoted string to indicate the plot's text font. By default, \code{"Avenir"}.}
}
\value{
A ggplot2 object.
}
\description{
This function will reformat a plot created by ggplot2 with basic CCMH theming. The theming relates to formatting text (i.e., font type, size) and background.
}
