% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMHr-package.R
\docType{package}
\name{CCMHr-package}
\alias{CCMHr}
\alias{CCMHr-package}
\title{CCMHr: CCMH Data Cleaning}
\description{
Functions to standardize routine data cleaning/processing procedures and the creation of visualizations.
}
\author{
\strong{Maintainer}: Rebecca Janis \email{ruj134@psu.edu}

Authors:
\itemize{
  \item Rebecca Janis \email{ruj134@psu.edu}
  \item Tyler White \email{tdw5466@psu.edu}
}

}
\keyword{internal}
