% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colum_classes.R
\name{check_column_classes}
\alias{check_column_classes}
\title{Check column class equality across two data frames}
\usage{
check_column_classes(dataframe1 = survey, dataframe2 = appts)
}
\arguments{
\item{dataframe1}{A data frame.}

\item{dataframe2}{Another data frame.}
}
\value{
A data frame with columns for each data frame and rows for each column in the data frames that has unequal column classes.
The observations indicate the column classes in each data frame.
}
\description{
A function to check that columns named the same across two data frames have the same classes.
To be used before merging two data frames to ensure that the columns with the same names will merge cleanly.
}
