% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_SDS.R
\name{check_SDS}
\alias{check_SDS}
\title{Check SDS item values}
\usage{
check_SDS(dat)
}
\arguments{
\item{dat}{The data frame containing the new year of (combined) data.}
}
\value{
A list with two elements. The first, \code{violations}, is a boolean indicating
whether or not any violations were found. The second, \code{SDS_viol}, is a vector of
strings containing the names of the SDS variables that have at least one value out of
bounds.
}
\description{
A function used to check the validity of the SDS items. Each SDS item has
a specific range of accepted values, and takes NA for missing values. This
function checks whether these constraints are satisfied, and if not, returns a list
containing the SDS item(s) which violated its/their constraints.
}
\note{
\code{SDSvalue_check} ...
}
