% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI_CCMH_export_support_functions.r
\name{clean_TI_export_appointment}
\alias{clean_TI_export_appointment}
\title{Clean TI appointment export file}
\usage{
clean_TI_export_appointment(
  data,
  key = CCMHr::TI_CCMH_export_appointment_key,
  save = FALSE,
  path = "TI_appointment_clean.csv"
)
}
\value{
A data frame with cleaned TI exported appointment data.
}
\description{
@param data A unquoted (when data file is an r object) or quoted string (when data file is located in a local directory) to specify the data file. If specified from directory, the data file should be a csv file.
@param key A data frame with two columns, orginal_name and new_name, to rename the columns in the appointment data. Default is \code{CCMHr::TI_CCMH_export_appointment_key}
@param save A logical statement indicates whether the data frame should be saved as a file under a local folder. If false, the data will be returned as an object. By default, \code{FALSE}
@param path A quoted string to indicate the path where the appointment data should be saved. By default, \code{TI_appointment_clean.csv}
}
