% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clicc_key.R
\docType{data}
\name{clicc_key}
\alias{clicc_key}
\title{CLICC key}
\format{
A data frame with columns:
\describe{
\item{Item}{CLICC item corresponding to the column name of each item in the data}
\item{Concern}{Name of the concern}
\item{Number}{CLICC number corresponding to the top concern in the CLICC_03 item}
}
}
\usage{
clicc_key
}
\description{
A key for the CLICC form items and numbers. Variables include Item (the column name of each CLICC item in the CCMH research data frame), Name (the name/label of each concern), and Number (the CLICC number corresponding to the top concern in the CLICC_03 item).
}
\keyword{datasets}
