% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_case_closure.r
\name{collapse_case_closure}
\alias{collapse_case_closure}
\title{Collapse data on case closure (select all that apply) across clients}
\usage{
collapse_case_closure(data, single.course = TRUE)
}
\arguments{
\item{data}{A data frame that contains data on Case Closure.}

\item{single.course}{A logical statement indicating whether data pertains to a single therapy course. If \code{TRUE}, the function assumes that each client has only one therapy course. If \code{FALSE}, the function will account for multiple therapy courses per client. By default, \code{TRUE}.}
}
\value{
Returns a data frame where data on Case Closure is collapsed across clients.
}
\description{
The case closure form could be filled out multiple times. This function ensures that if a client has multiple case closure records, all unique case closure reasons and events are captured in a single row for that client. This only applies to the select all that apply questions.
}
