% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_creation.R
\name{create_course}
\alias{create_course}
\title{Create courses of therapy within data}
\usage{
create_course(data, firstOnly = FALSE)
}
\arguments{
\item{data}{CCMH data files with UniqueClientID and Date variables}

\item{firstOnly}{If FALSE, will return the whole data set. If TRUE, will return only each client's first course.}
}
\value{
A data frame with added variables:
`UniqueClientID_byCourse`: an ID unique to each course of therapy. This is created by adding .1 to UniqueClientID for their first course (e.g. 100.1), .2 for their second course (100.2), etc.
`RankCourse`: The course number that the line of data belongs to, numbered 1 through N of courses.
`FirstCourse`: A dichotomous variable indicating whether the lines of data belong to a client's first course or not.
}
\description{
Create courses of therapy in CCMH data.
}
