% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_year.r
\name{create_Data_year}
\alias{create_Data_year}
\title{Create Data_year variable}
\usage{
create_Data_year(data)
}
\arguments{
\item{data}{A data file.}
}
\value{
A data frame with a new variable Data_year.
}
\description{
Create Data_year variable
}
