% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicate_appointments.R
\name{delete_duplicate_appointments}
\alias{delete_duplicate_appointments}
\title{Delete duplicate appointments.}
\usage{
delete_duplicate_appointments(
  data,
  client_identifier = "UniqueClientID",
  appointment_identifier = "AppointID"
)
}
\arguments{
\item{data}{A data frame containing rows related to appointments.}

\item{client_identifier}{A quoted string to indicate the column uniquely identifying each client. By default, \code{"UniqueClientID"}.}

\item{appointment_identifier}{A quoted string to indicate the column uniquely identifying each appointment. By default, \code{AppointID}. If \code{"AppointID"} is not available, use \code{"Date"} instead.}
}
\value{
A data frame with only the first instance of each appointment.
}
\description{
For reasons not discussed here, data on appointments may have duplicates (e.g., multiple rows with the same type of data). This function removes duplicate data from appointments.
}
