% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetches a data set from a CCMH - Documents repository}
\usage{
fetch_data(data.year)
}
\arguments{
\item{data.year}{An unquoted integer, unquoted number, or "all". The unquoted integer or number will fetch data from a single data year (i.e., data from 2010 to the most recent data year). If "all", the complete data set with all data years is fetched from a parquet file.}
}
\value{
If data.year equals an unquoted integer or number, a data frame containing a single data year will be returned. If data.year equals "all", an arrow table containing data from all data years is returned.
}
\description{
Fetches a data set from a CCMH - Documents repository. Each single year data frame can be fetched, along with the complete data frame. To run this function, the programmer must set the working directory within the CCMH - Documents folder. Typically, only CCMH staff have access to the CCMH - Documents folder.
}
