% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_select.r
\name{filter_select}
\alias{filter_select}
\title{Filter rows and select columns specific to the chosen data years and forms.}
\usage{
filter_select(data, data.year = NULL, forms = NULL)
}
\arguments{
\item{data}{A data frame or arrow table that contains CCMH data.}

\item{data.year}{A numeric value or list of numeric values that specifies the data years needed in the returned data frame. If \code{NULL}, all data years are included. By default, \code{NULL}.}

\item{forms}{A string or list of strings that specifies the forms needed in the returned data frame. Options include: \code{"ccaps"}, \code{"sds"}, \code{"clicc"}, \code{"closure"}, \code{"appointment"}, \code{"audit"}, \code{"provider"}, \code{"dx"}, \code{NULL}. If \code{NULL}, all forms are included. By default, \code{NULL}.}
}
\value{
Outputs a data frame with rows and columns about specified data years and forms. Please note that if the input is an arrow table, the output will be a data frame.
}
\description{
Filter rows to include only rows about specified data years and forms, and select only columns relevant to those years and forms. This simplifies and reduces the size of large datasets by removing unnecessary data.
}
