% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_year.r
\name{format_Data_year}
\alias{format_Data_year}
\title{Format Data_year variable}
\usage{
format_Data_year(data)
}
\arguments{
\item{data}{A data file.}
}
\value{
A data frame with new variables: Data_year_long, Data_year_formal, and Data_year_formalshort.
}
\description{
Format Data_year variable
}
