% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_subscales.R
\name{format_subscales}
\alias{format_subscales}
\title{Format subscale names}
\usage{
format_subscales(data, col, formal = F)
}
\arguments{
\item{data}{Data frame}

\item{col}{Column with subscale names}

\item{formal}{If FALSE (default), returns informal versions of names (e.g. Anxiety, Eating). If TRUE, returns formal versions of names (e.g. Generalized Anxiety, Eating Concerns).}
}
\value{
A data frame with renamed subscales.
}
\description{
Renames a column of CCAPS subscales to remove 34 or 62 and underscores in subscale name for use in making plots, tables, etc.
}
\examples{
\dontrun{
format_subscales(data, subscale, formal = F)
}

}
