% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqtab.R
\name{freqtab}
\alias{freqtab}
\title{Create frequency table.}
\usage{
freqtab(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A data frame or table containing details from the frequency table.
}
\description{
The function creates a frequency table based on the values within a vector. The frequency table outputs the percentage and cumulative percentage based on the vector values.
}
