% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Extract the mode value from a vector.}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode value(s) within a vector.
}
\description{
The function extracts the mode value(s) from a vector. \code{NA} responses are omitted.
}
