% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouply.R
\name{grouply}
\alias{grouply}
\title{grouply}
\usage{
grouply(f, ...)
}
\arguments{
\item{f}{A function}

\item{...}{A grouping variable}
}
\description{
grouply
}
\examples{
mtcars \%>\% grouply(dplyr::mutate, cyl)(M = mean(wt))

}
