% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_primary_therapist.r
\name{identity_primary_therapist}
\alias{identity_primary_therapist}
\title{Identity primary therapist}
\usage{
identity_primary_therapist(
  data,
  percent.threshold = 0.5,
  primary.therapist.only = FALSE
)
}
\arguments{
\item{data}{A data file.}

\item{percent.threshold}{A numeric value ranging between .50 (50\%) to 1 (\%100) to indicate the percentage of appointments across the UniqueClientID, CcmhID, and Data_year required to classify a therapist as the primary therapist. For example, 0.75 would contend that atleast 75\% of appointments should be conducted by a therapist to classify them as the primary therapist. Default is \code{0.50}.}

\item{primary.therapist.only}{A logical statement to indicate if only the primary therapist should be returned. Default is \code{FALSE}.}
}
\description{
Identity primary therapist
}
