% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_present.R
\name{last_present}
\alias{last_present}
\title{Extract the last value from a vector.}
\usage{
last_present(x, order_by = NULL, default = NA)
}
\arguments{
\item{x}{A vector.}

\item{order_by}{Another vector to sort \code{x} by.}

\item{default}{A numeric value to use if the \code{x} position does not exist in the input.}
}
\value{
The last non-NA value in a vector.
}
\description{
The function extracts the last non-NA value from a vector.
}
