% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRDa.R
\name{loadRDa}
\alias{loadRDa}
\title{load .RDa file into new object name}
\usage{
loadRDa(fileName)
}
\arguments{
\item{fileName}{The file path to a .rda or .rdata file to be loaded into a new object name}
}
\description{
Loads .rda and .rdata files with the ability to assign them to a new object name instead of the one they were originally saved with.
}
\examples{
\dontrun{
new_name <- loadRDa("filepath/file.rda")
}

}
