% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function.R
\name{new_function}
\alias{new_function}
\title{A New Function}
\usage{
new_function(n)
}
\arguments{
\item{n}{The number of standard normal random variates you want to generate.}
}
\value{
A pretty vanilla looking histogram of standard normal random variables.
}
\description{
Takes an integer-valued argument and generates that many standard normal random variables, then plots a histogram of these random variates.
}
\details{
Just checking to see if github is properly linked to this project
}
\keyword{Histogram,}
\keyword{Normal}
\keyword{Random}
\keyword{Variable}
