% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
A pipe operator that allows for the use of the magrittr pipe operator (\%>\%). This pipe operator is used to chain together multiple functions in a more readable way. CCMHr uses |> as the pipe operator, which is the native pipe operator in R 4.1.0 and later. The utils-pipe function allows for the use of the magrittr pipe operator (\%>\%) in earlier versions of R.
}
\keyword{internal}
