% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\alias{line_plot}
\title{Creates a line plot graph.}
\usage{
plot_line(
  data,
  x.var,
  y.var,
  group.var = "",
  color = CCMHr::CCMH_categorical_palettes$tolmuted_category10,
  save = FALSE,
  path = "plot.png",
  plot.width = 12,
  plot.height = 9,
  plot.units = "in",
  plot.dpi = 360,
  plot.device = NULL,
  plot.scale = 1,
  hide.group.items = NULL,
  plot.title = NULL,
  x.title = NULL,
  y.title = NULL,
  plot.title.size = 20,
  axis.title.size = 16,
  axis.label.size = 14,
  text.font = "Avenir",
  y.label.type = "numeric",
  y.min = NULL,
  y.max = NULL,
  y.breaks = NULL,
  y.expand = ggplot2::waiver(),
  caption = FALSE,
  caption.size = 12,
  caption.vjust = -3,
  legend.position = "none",
  legend.order.manual = NULL,
  legend.title = NULL,
  legend.title.size = 16,
  legend.label.size = 14,
  coord.flip = FALSE,
  y.grid.major = FALSE,
  y.grid.major.color = "#d3d3d3",
  y.grid.major.size = 0.5,
  y.grid.major.linetype = 1,
  y.grid.minor = FALSE,
  y.grid.minor.color = "#d3d3d3",
  y.grid.minor.size = 0.25,
  y.grid.minor.linetype = 1,
  y.axis.line = TRUE,
  x.axis.line = TRUE,
  remove.axis.ticks = FALSE,
  line.size = 1.5,
  line.type = "solid",
  point = FALSE,
  show.point.legend = FALSE,
  point.size = 2,
  point.shape = 19,
  plot.element1 = NULL,
  plot.element2 = NULL,
  plot.element3 = NULL,
  plot.element4 = NULL,
  plot.element5 = NULL,
  plot.element6 = NULL,
  plot.element7 = NULL,
  plot.element8 = NULL,
  plot.element9 = NULL
)

line_plot(
  data,
  x.var,
  y.var,
  group.var = "",
  color = CCMHr::CCMH_categorical_palettes$tolmuted_category10,
  save = FALSE,
  path = "plot.png",
  plot.width = 12,
  plot.height = 9,
  plot.units = "in",
  plot.dpi = 360,
  plot.device = NULL,
  plot.scale = 1,
  hide.group.items = NULL,
  plot.title = NULL,
  x.title = NULL,
  y.title = NULL,
  plot.title.size = 20,
  axis.title.size = 16,
  axis.label.size = 14,
  text.font = "Avenir",
  y.label.type = "numeric",
  y.min = NULL,
  y.max = NULL,
  y.breaks = NULL,
  y.expand = ggplot2::waiver(),
  caption = FALSE,
  caption.size = 12,
  caption.vjust = -3,
  legend.position = "none",
  legend.order.manual = NULL,
  legend.title = NULL,
  legend.title.size = 16,
  legend.label.size = 14,
  coord.flip = FALSE,
  y.grid.major = FALSE,
  y.grid.major.color = "#d3d3d3",
  y.grid.major.size = 0.5,
  y.grid.major.linetype = 1,
  y.grid.minor = FALSE,
  y.grid.minor.color = "#d3d3d3",
  y.grid.minor.size = 0.25,
  y.grid.minor.linetype = 1,
  y.axis.line = TRUE,
  x.axis.line = TRUE,
  remove.axis.ticks = FALSE,
  line.size = 1.5,
  line.type = "solid",
  point = FALSE,
  show.point.legend = FALSE,
  point.size = 2,
  point.shape = 19,
  plot.element1 = NULL,
  plot.element2 = NULL,
  plot.element3 = NULL,
  plot.element4 = NULL,
  plot.element5 = NULL,
  plot.element6 = NULL,
  plot.element7 = NULL,
  plot.element8 = NULL,
  plot.element9 = NULL
)
}
\arguments{
\item{data}{A data file containing variables that will be used in the plot.}

\item{x.var}{A quoted or unquoted string indicates the variable that will be plotted on the x-axis. The variable class must be a character or factor.}

\item{y.var}{A quoted or unquoted string indicates the variable that will be plotted on the y-axis. The variable class must be numeric.}

\item{group.var}{A quoted string to indicate the variable that will group data/lines. By default, \code{""}.}

\item{color}{A hex code or list of hex codes that indicates the color(s) of the lines. If a variable groups the data, the argument specifies the color of the lines or items within the grouped variable. For example, the group variable items could be the names of automobile companies (i.e., Ford, Dodge, BMW). You can specify a random color for each company using this list: \code{c("#21501b", "#c51329", "#074e67")}. Alternatively, you can specify the color directly for each company using this list: \code{c("Ford" = "#21501b", "Dodge" = "#c51329", "BMW" = "#074e67")}. The number of hex codes must be equal to or greater than the number of responses in the grouped data. If data is not grouped, all the lines will have the same color. The default colors are color-blind safe. By default, \code{CCMHr::CCMH_categorical_palettes$tolmuted_category10}.}

\item{save}{A logical argument that indicates whether the plot should be saved as a file under a local folder. If \code{FALSE}, the plot will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway, name, and type if \code{save = TRUE}. By default, \code{"plot.png"}.}

\item{plot.width}{A numeric value to indicate the plot's width. By default, \code{12}.}

\item{plot.height}{A numeric value to indicate the plot's height. By default, \code{9}.}

\item{plot.units}{A quoted string to indicate the plot's width and height in size units. Options include \verb{in}, \code{cm}, \code{mm}, and \code{px}. By default, \verb{in}.}

\item{plot.dpi}{A numeric value to indicate the plot's image resolution. By default, \code{360}.}

\item{plot.device}{A quoted string or function to indicate the plot's device. If saving the plot as a PDF, use \code{cairo_pdf}. By default, \code{NULL}.}

\item{plot.scale}{A numeric value to indicate the plot's scale. By default, \code{1}.}

\item{hide.group.items}{A quoted string or list to indicate the group variable items to be hidden in the plot. The argument hides the listed grouped item bars in the plot while maintaining the position of the hidden bars. This argument is intended for use when preparing plots for presentation, where the presenter wants to present data in a specified order instead of displaying all the data at once. By default, \code{NULL}.}

\item{plot.title}{A quoted string to indicate the title of the plot. By default, \code{NULL}.}

\item{x.title}{A quoted string to indicate the x-axis title. By default, \code{NULL}.}

\item{y.title}{A quoted string to indicate the y-axis title. By default, \code{NULL}.}

\item{plot.title.size}{A numeric value to indicate the plot title text size. By default, \code{20}.}

\item{axis.title.size}{A numeric value to indicate the axis title text size. By default, \code{16}.}

\item{axis.label.size}{A numeric value to indicate the axis labels text size. By default, \code{14}.}

\item{text.font}{A quoted string to indicate the plot's text font. By default, \code{"Avenir"}.}

\item{y.label.type}{A quoted string or list of quoted strings to specify the format of the y-axis labels. Options include \code{"numeric"}, \code{"percent"}, \code{"comma"}, \code{"sci"}, or \code{"dollar"}. By default, \code{"numeric"}.}

\item{y.min}{A numeric value to indicate the minimum number presented on the y-axis. By default, \code{NULL}.}

\item{y.max}{A numeric value to indicate the maximum number presented on the y-axis. By default, \code{NULL}.}

\item{y.breaks}{Numeric values or a sequence of numeric values that indicate breaks on the y-axis. By default, \code{NULL}.}

\item{y.expand}{A list of two numeric values to indicate the expansion of the y-axis. By default, \code{ggplot2::waiver()}.}

\item{caption}{A logical argument to indicate whether the CCMH caption should be displayed on the bottom right of the plot. If \code{TRUE}, the caption "Source: The Center for Collegiate Mental Health" is displayed on the bottom right of the plot. By default, \code{FALSE}.}

\item{caption.size}{A numeric value to indicate the caption text size. By default, \code{12}.}

\item{caption.vjust}{A numeric value to indicate the caption vertical adjustment. By default, \code{-3}.}

\item{legend.position}{A quoted string or numeric vector to indicate the location of the legend. Options include \code{"left"},\code{"top"}, \code{"right"}, \code{"bottom"}, \code{"none"}, or numeric vector c(x,y). By default, \code{"none"}.}

\item{legend.order.manual}{A list of quoted strings to indicate the order of the legend. By default, \code{NULL}.}

\item{legend.title}{A quoted string to indicate the legend title. By default, \code{NULL}.}

\item{legend.title.size}{A numeric value to indicate the legend title text size. By default, \code{16}.}

\item{legend.label.size}{A numeric value to indicate the legend label text size. By default, \code{14}.}

\item{coord.flip}{A logical statement to indicate if the plot's x and y axes should be flipped. By default, \code{FALSE}.}

\item{y.grid.major}{A logical argument to indicate whether the y-axis major grid lines should be added to the plot. If \code{TRUE}, the y-axis major grid lines are displayed on the plot. By default, \code{FALSE}.}

\item{y.grid.major.color}{A hex code that indicates the color of the y-axis major grid lines. By default, \code{"#d3d3d3"} or a shade of light grey.}

\item{y.grid.major.size}{A numeric value to indicate the thickness of the y-axis major grid lines. By default, \code{0.5}.}

\item{y.grid.major.linetype}{A numeric value to indicate the line type of the y-axis major grid. By default, \code{1}.}

\item{y.grid.minor}{A logical argument to indicate if the y-axis minor grid lines should be added to the plot. If \code{TRUE}, the y-axis minor grid lines are displayed on the plot. By default, \code{FALSE}.}

\item{y.grid.minor.color}{A hex code that indicates the color of the y-axis minor grid lines. By default, \code{"#d3d3d3"} or a shade of light grey.}

\item{y.grid.minor.size}{A numeric value to indicate the thickness of the y-axis minor grid lines. By default, \code{0.25}.}

\item{y.grid.minor.linetype}{A numeric value to indicate the line type of the y-axis minor grid. By default, \code{1}.}

\item{y.axis.line}{A logical argument to indicate whether the y-axis line should be included. If \code{FALSE}, the y-axis line will not be displayed. By default, \code{TRUE}.}

\item{x.axis.line}{A logical argument to indicate whether the x-axis line should be included. If \code{FALSE}, the x-axis line will not be displayed. By default, \code{TRUE}.}

\item{remove.axis.ticks}{A logical argument to indicate whether the axis ticks should be excluded. If \code{TRUE}, the axis ticks will not be displayed on the plot. By default, \code{FALSE}.}

\item{line.size}{A numeric value to indicate the line thickness. By default, \code{1.5}.}

\item{line.type}{A quoted string to indicate the line type. By default, \code{"solid"}.}

\item{point}{A logical argument to indicate whether the plot should include points on each. If \code{TRUE}, points will be added to each line. By default, \code{FALSE}.}

\item{show.point.legend}{A logical argument to indicate whether the plot should include points on each. If \code{TRUE}, points will be added to each line. By default, \code{FALSE}.}

\item{point.size}{A numeric value to indicate point size. By default, \code{2}.}

\item{point.shape}{A numeric value to indicate point shape. By default, \code{19}.}

\item{plot.element1}{A ggplot2 plot function and arguments needed to add graphical element(s) needed to complete a specific task, but are not specified as an argument in the function. For example, the axis label text color will always be black unless specified in one of the plot.element arguments (i.e., plot.element1 to plot.element9). To change the axis label text color, one of the plot.element arguments should be specified as follows: \code{plot.element1 = ggplot2::theme(axis.text = ggplot2::element_text(color = "green"))}. By default, \code{NULL}.}

\item{plot.element2}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element3}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element4}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element5}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element6}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element7}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element8}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}

\item{plot.element9}{A ggplot2 plot function and arguments needed to add graphical element(s) required to complete a specific task, but are not specified as an argument in the function. By default, \code{NULL}.}
}
\value{
A line plot is returned as an object or saved as a file in a local directory.
}
\description{
This function creates a line plot graph.
}
