% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\alias{scatter_plot}
\title{A function to create standardized CCMH scatter plot's}
\usage{
plot_scatter(
  data,
  x.var,
  y.var,
  group.var = "",
  color = c("#666666", "#1B9E77", "#E6AB02", "#D95F02", "#6db6ff", "#b66dff", "#ff6db6",
    "#920000"),
  save = FALSE,
  path = "plot.png",
  plot.width = 15.24,
  plot.height = 12.7,
  plot.units = "cm",
  plot.dpi = 320,
  plot.device = NULL,
  hide.group.items = NULL,
  plot.title = NULL,
  x.title = NULL,
  y.title = NULL,
  plot.title.size = 20,
  axis.title.size = 16,
  axis.label.size = 14,
  text.font = "Avenir",
  y.label.type = "numeric",
  y.min = NULL,
  y.max = NULL,
  y.breaks = NULL,
  x.label.type = "numeric",
  x.min = NULL,
  x.max = NULL,
  x.breaks = NULL,
  caption = FALSE,
  caption.size = 12,
  caption.vjust = 0,
  legend.position = "none",
  legend.order.manual = NULL,
  legend.title = NULL,
  legend.title.size = 16,
  legend.label.size = 14,
  coord.flip = FALSE,
  y.grid.major = FALSE,
  y.grid.major.color = "#d3d3d3",
  y.grid.major.size = 0.5,
  y.grid.major.linetype = 1,
  y.grid.minor = FALSE,
  y.grid.minor.color = "#d3d3d3",
  y.grid.minor.size = 0.25,
  y.grid.minor.linetype = 1,
  x.grid.major = FALSE,
  x.grid.major.color = "#d3d3d3",
  x.grid.major.size = 0.5,
  x.grid.major.linetype = 1,
  x.grid.minor = FALSE,
  x.grid.minor.color = "#d3d3d3",
  x.grid.minor.size = 0.25,
  x.grid.minor.linetype = 1,
  y.axis.line = TRUE,
  x.axis.line = TRUE,
  remove.axis.ticks = FALSE,
  fit.line = FALSE,
  fit.line.type = NULL,
  fit.line.ci = FALSE,
  fit.line.fullrange = FALSE,
  point_size = 1.5,
  point_shape = 19,
  plot.element1 = NULL,
  plot.element2 = NULL,
  plot.element3 = NULL,
  plot.element4 = NULL,
  plot.element5 = NULL,
  plot.element6 = NULL,
  plot.element7 = NULL,
  plot.element8 = NULL,
  plot.element9 = NULL
)

scatter_plot(
  data,
  x.var,
  y.var,
  group.var = "",
  color = c("#666666", "#1B9E77", "#E6AB02", "#D95F02", "#6db6ff", "#b66dff", "#ff6db6",
    "#920000"),
  save = FALSE,
  path = "plot.png",
  plot.width = 15.24,
  plot.height = 12.7,
  plot.units = "cm",
  plot.dpi = 320,
  plot.device = NULL,
  hide.group.items = NULL,
  plot.title = NULL,
  x.title = NULL,
  y.title = NULL,
  plot.title.size = 20,
  axis.title.size = 16,
  axis.label.size = 14,
  text.font = "Avenir",
  y.label.type = "numeric",
  y.min = NULL,
  y.max = NULL,
  y.breaks = NULL,
  x.label.type = "numeric",
  x.min = NULL,
  x.max = NULL,
  x.breaks = NULL,
  caption = FALSE,
  caption.size = 12,
  caption.vjust = 0,
  legend.position = "none",
  legend.order.manual = NULL,
  legend.title = NULL,
  legend.title.size = 16,
  legend.label.size = 14,
  coord.flip = FALSE,
  y.grid.major = FALSE,
  y.grid.major.color = "#d3d3d3",
  y.grid.major.size = 0.5,
  y.grid.major.linetype = 1,
  y.grid.minor = FALSE,
  y.grid.minor.color = "#d3d3d3",
  y.grid.minor.size = 0.25,
  y.grid.minor.linetype = 1,
  x.grid.major = FALSE,
  x.grid.major.color = "#d3d3d3",
  x.grid.major.size = 0.5,
  x.grid.major.linetype = 1,
  x.grid.minor = FALSE,
  x.grid.minor.color = "#d3d3d3",
  x.grid.minor.size = 0.25,
  x.grid.minor.linetype = 1,
  y.axis.line = TRUE,
  x.axis.line = TRUE,
  remove.axis.ticks = FALSE,
  fit.line = FALSE,
  fit.line.type = NULL,
  fit.line.ci = FALSE,
  fit.line.fullrange = FALSE,
  point_size = 1.5,
  point_shape = 19,
  plot.element1 = NULL,
  plot.element2 = NULL,
  plot.element3 = NULL,
  plot.element4 = NULL,
  plot.element5 = NULL,
  plot.element6 = NULL,
  plot.element7 = NULL,
  plot.element8 = NULL,
  plot.element9 = NULL
)
}
\arguments{
\item{data}{A data file containing all variables that will be used in the plot}

\item{x.var}{A quoted string or unquoted characters to indicate the variable that will be plotted on the x-axis. The variable class must be numeric.}

\item{y.var}{A quoted string or unquoted characters to indicate the variable that will be plotted on the y axis. The variable class must be numeric.}

\item{group.var}{A quoted string to indicate the variable that will group the points on the x and y-axis and specify the color of the points. By default, \code{""}.}

\item{color}{A hex code or list of hex codes that indicates the color of the grouped points. See Note 1 for more details. By default, \code{c('#E6AB02', '#1B9E77', '#666666', '#D95F02', '#6db6ff', '#b66dff', '#ff6db6', '#920000')}.}

\item{save}{A logical statement indicates whether the plot should be saved as a file under a local folder. If false, the plot will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway and name if \code{save = TRUE}. By default, \code{"plot.png"}.}

\item{plot.width}{A numeric value to indicate the plot's width. By default, \code{15.24}.}

\item{plot.height}{A numeric value to indicate the plot's height. By default, \code{12.7}.}

\item{plot.units}{A quoted string to indicate the plot's width and height size units. Options include \verb{in}, \code{cm}, \code{mm}, and \code{px}. By default, \code{cm}.}

\item{plot.dpi}{A numeric value to indicate the plot's image resolution. By default, \code{320}.}

\item{plot.device}{A quoted string or function to indicate the plot's device. By default, \code{NULL}.}

\item{hide.group.items}{A quoted string or list to indicate the group variable items to be hidden in the plot. See Note 2. By default, \code{NULL}.}

\item{plot.title}{A quoted string to indicate the title of the plot. By default, \code{NULL}.}

\item{x.title}{A quoted string to create the x axis title. By default, \code{NULL}.}

\item{y.title}{A quoted string to create the y axis title. By default, \code{NULL}.}

\item{plot.title.size}{A numeric value to indicate the plot title text size. By default, \code{20}.}

\item{axis.title.size}{A numeric value to indicate the axis title text size. By default, \code{16}.}

\item{axis.label.size}{A numeric value to indicate the axis labels text size. By default, \code{14}.}

\item{text.font}{A quoted string to indicate the plot's text font. By default, \code{Avenir}.}

\item{y.label.type}{A quoted string or list to customize the y axis labels. Options include \code{"numeric"}, \code{"percent"}, \code{"comma"}, \verb{"sci}, or \code{"dollar"} . By default, \code{numeric}.}

\item{y.min}{A numeric value to indicate the minimum number presented on the y axis. By default, \code{NULL}.}

\item{y.max}{A numeric value to indicate the maximum number presented on the y axis. By default, \code{NULL}.}

\item{y.breaks}{Numeric values or a sequence that indicates breaks on the y axis. By default, \code{NULL}.}

\item{x.label.type}{A quoted string or list to customize the x axis labels. Options include \code{"numeric"}, \code{"percent"}, \code{"comma"}, \verb{"sci}, or \code{"dollar"} . By default, \code{numeric}.}

\item{x.min}{A numeric value to indicate the minimum number presented on the x axis. By default, \code{NULL}.}

\item{x.max}{A numeric value to indicate the maximum number presented on the x axis. By default, \code{NULL}.}

\item{x.breaks}{Numeric values or a sequence that indicates breaks on the x axis. By default, \code{NULL}.}

\item{caption}{A logical statement to indicate whether the CCMH caption should be included in the plot. By default, \code{FALSE}.}

\item{caption.size}{A numeric value to indicate the caption text size. By default, \code{12}.}

\item{caption.vjust}{A numeric value to indicate the caption vertical adjustment. By default, \code{0}.}

\item{legend.position}{A quoted string or numeric vector to indicate the location of the legend. Options include \code{"left"},\code{"top"}, \code{"right"}, \code{"bottom"}, \code{"none"}, or numeric vector c(x,y). By default, \code{"none"}.}

\item{legend.order.manual}{A list of quoted strings to indicate the order of the legend. By default, \code{NULL}.}

\item{legend.title}{A quoted string to indicate the legend title. By default, \code{NULL}.}

\item{legend.title.size}{A numeric value to indicate the legend title text size. By default, \code{16}.}

\item{legend.label.size}{A numeric value to indicate the legend label text size. By default, \code{14}.}

\item{coord.flip}{A logical statement to indicate if the plot's x and y axis should be flipped. By default, \code{FALSE}.}

\item{y.grid.major}{A logical statement to indicate if the y axis major grid lines should be added to the plot. By default, \code{FALSE},}

\item{y.grid.major.color}{A hex code that indicates the color of the y axis major grid lines. By default, \code{"#d3d3d3"} or a shade of light grey.}

\item{y.grid.major.size}{A numeric value to indicate the thickness of the y axis major grid lines. By default, \code{0.5}.}

\item{y.grid.major.linetype}{A numeric value to indicate the line type of the y axis major grid. By default, \code{1}.}

\item{y.grid.minor}{A logical statement to indicate if the y axis minor grid lines should be added to the plot. By default, \code{FALSE},}

\item{y.grid.minor.color}{A hex code that indicates the color of the y axis minor grid lines. By default, \code{"#d3d3d3"} or a shade of light grey.}

\item{y.grid.minor.size}{A numeric value to indicate the thickness of the y axis minor grid lines. By default, \code{0.25}.}

\item{y.grid.minor.linetype}{A numeric value to indicate the line type of the y axis minor grid. By default, \code{1}.}

\item{x.grid.major}{A logical statement to indicate if the y axis major grid lines should be added to the plot. By default, \code{FALSE},}

\item{x.grid.major.color}{A hex code that indicates the color of the y axis major grid lines. By default, \code{"#d3d3d3"} or a shade of light grey.}

\item{x.grid.major.size}{A numeric value to indicate the thickness of the y axis major grid lines. By default, \code{0.5}.}

\item{x.grid.major.linetype}{A numeric value to indicate the line type of the y axis major grid. By default, \code{1}.}

\item{x.grid.minor}{A logical statement to indicate if the x axis minor grid lines should be added to the plot. By default, \code{FALSE},}

\item{x.grid.minor.color}{A hex code that indicates the color of the x axis minor grid lines. By default, \code{"#d3d3d3"} or a shade of light grey.}

\item{x.grid.minor.size}{A numeric value to indicate the thickness of the x axis minor grid lines. By default, \code{0.25}.}

\item{x.grid.minor.linetype}{A numeric value to indicate the line type of the x axis minor grid. By default, \code{1}.}

\item{y.axis.line}{A logical statement to indicate if the y axis line should be included. By default, \code{TRUE}.}

\item{x.axis.line}{A logical statement to indicate if the x axis line should be included. By default, \code{TRUE}.}

\item{remove.axis.ticks}{A logical statement to indicate if the axis ticks should be excluded. By default, \code{FALSE}.}

\item{fit.line}{A logical statement to indicate if the plot should have a fitted line. By default, \code{FALSE}.}

\item{fit.line.type}{A quoted string to indicate the method of fitting a line. See information on geom_smooth for more information. By default, \code{NULL}.}

\item{fit.line.ci}{A logical statement to indicate if the plotted line should have confidence intervals. By default, \code{FALSE}.}

\item{fit.line.fullrange}{A logical statement to indicate if the plotted line should cross the entire plot. By default, \code{FALSE}.}

\item{point_size}{A numeric value to indicate the point size. By default, \code{1.5}.}

\item{point_shape}{A numeric value to indicate the point shape. By default, \code{19}.}

\item{plot.element1}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element2}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element3}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element4}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element5}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element6}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element7}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element8}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element9}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}
}
\value{
A column plot returned as a object or saved under a local file.
}
\description{
Creates standardized CCMH scatter plot's using ggplot2.
}
\note{
Note 1. The argument \code{color} specifies the color of points within the grouped variable. For example, the group variable items could be the name of automobile companies (i.e., Ford, Dodge, BMV). You can specify color randomly for each company using this list: \code{c("#21501b", "#c51329", "#074e67")}. Or you can specify color directly to each specific company using this list: \code{c("Ford" = "#21501b", "Dodge" = "#c51329", "BMV" = "#074e67")}. Default colors are color blind safe.

Note 2. The \code{hide.group.items} function hides points in the plot while keeping the hidden points position. This argument was intended to be used when preparing plots for presentation, where the presenter wants to present data in a specified order instead of displaying all the data at once. Fit lines will not be removed.

The \code{plot.element} arguments specify additional ggplot2 graphical element(s) needed to complete a specific task, but are not specified as an argument in the \code{scatter_plot} function. For example, the axis label text will always be black unless specified in one of the \code{plot.element} arguments. To change the axis label text color in one of the \code{plot.element} arguments, you would first create an object to represent graphing element (e.g., green.axis.text <- ggplot2::theme(axis.text = ggplot2::element_text(color = "green"))). Then the object would be added to one of the \code{plot.element} arguments (e.g., \code{plot.element01 = green.axis.text}).
}
