% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_race_ethnicity.r
\name{recode_race_ethnicity}
\alias{recode_race_ethnicity}
\title{Recode the multi-select responses for race/ethnicity into a single variable.}
\usage{
recode_race_ethnicity(data)
}
\arguments{
\item{data}{A data file containing race/ethnicity variables (e.g., SDS_1095_01).}
}
\value{
Outputs the original dataset along with a new variable, SDS_1095, that recodes multi-select responses for race/ethnicity into a single variable.
}
\description{
Creates a single variable that recodes multi-select responses for race/ethnicity into a single variable. Those who selected multiple races/ethnicities are categorized as "Multi-racial".
}
