% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty.R
\name{remove_empty}
\alias{remove_empty}
\title{Remove empty variables}
\usage{
remove_empty(data)
}
\arguments{
\item{data}{A data object}
}
\value{
The data object without any variables that were completely empty
}
\description{
Used in data request cleaning to remove variables that don't contain any data
}
