% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_subscales.R
\name{rename_subscales_long}
\alias{rename_subscales_long}
\title{Rename CCAPS subscales within a long format.}
\usage{
rename_subscales_long(data, column, formal = FALSE)
}
\arguments{
\item{data}{A data frame that contains a column with CCAPS subscale variable names. The column must have valid CCAPS names such as "Depression34", "Anxiety34", "Social_Anxiety34", "Academics34", "Eating34", "Hostility34", "Alcohol34", "DI", "Depression62", "Eating62", "Substance62", "Anxiety62", "Hostility62", "Social_Anxiety62", "Family62", and "Academics62".}

\item{column}{An unquoted string to indicate the name of the column that contains the CCAPS subscale variable names.}

\item{formal}{A logical argument to determine whether CCAPS subscale names should be in a formal or informal format. If \code{FALSE}, the CCAPS subscale names are renamed in an informal format (i.e., "Depression", "Anxiety", "Social Anxiety", "Academics", "Eating", "Frustration/Anger", "Alcohol Use", "Substance Use", "Family Distress", "Distress Index"). If \code{TRUE}, the CCAPS subscale names are renamed in an informal format (i.e., "Depression", "Generalized Anxiety", "Social Anxiety", "Academic Distress", "Eating Concerns", "Frustration/Anger", "Alcohol Use", "Substance Use", "Family Distress", "Distress Index"). By default, \code{FALSE}.}
}
\value{
A data frame with the column containing CCAPS subscale names is recoded to the formal or informal names.
}
\description{
The CCAPS subscale variable names are "Depression34", "Anxiety34", "Social_Anxiety34", "Academics34", "Eating34", "Hostility34", "Alcohol34", "DI", "Depression62", "Eating62", "Substance62", "Anxiety62", "Hostility62", "Social_Anxiety62", "Family62", and "Academics62". When creating visualizations, we often rename the subscales to make them more salient (e.g., "DI" corresponds to "Distress Index"). This function renames the subscales when CCAPS subscale variable names are stored in a data frame in a long format (i.e., the subscale variables are stored in a single column, and multiple rows correspond to a single client, date, and subscale). CCAPS34 and CCAPS62 subscales variable names can be in the same column.
}
