% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_subscales.R
\name{rename_subscales_wide}
\alias{rename_subscales_wide}
\alias{rename_subscales_long}
\title{Rename CCAPS subscales}
\usage{
rename_subscales_wide(data, CCAPS62 = F)

rename_subscales_long(data, column, formal = F)
}
\arguments{
\item{data}{Data frame}

\item{CCAPS62}{If FALSE (default), CCAPS-34 subscale columns are renamed.}

\item{column}{Column with subscale names}

\item{formal}{If FALSE (default), returns informal versions of names (e.g. Anxiety, Eating). If TRUE, returns formal versions of names (e.g. Generalized Anxiety, Eating Concerns).}
}
\value{
A data frame with renamed subscale column names or single column of subscale names.
}
\description{
The \code{rename_subscales} family of functions renames the CCAPS subscales for better presentation (e.g. removing underscores, removing 34 or 62 from subscale names)
\code{rename_subscales_wide} takes a wide data set where each subscale is its own column.
\code{rename_subscales_long} takes a long data set where with a single column of CCAPS subscale names
}
\examples{
\dontrun{
format_subscales(data, subscale, formal = F)
}
}
