% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_items.R
\name{required_items}
\alias{required_items}
\title{Detects whether specified variables are present in the dataset.}
\usage{
required_items(
  data,
  items,
  message = "The following variable(s) need to be present in data:"
)
}
\arguments{
\item{data}{A data frame.}

\item{items}{A quoted string or list of quoted strings that specifies variable names required in a data frame.}

\item{message}{A quoted string of the returned error message when the specified variable(s) are missing from a data frame. By default, \code{"The following variable(s) need to be present in the data frame:"} and a list of the missing variables.}
}
\value{
When all the required variables are present, nothing is returned. When at least one required variable is not present, an error message is returned. The default error message is \code{"The following variable(s) need to be present in the data frame:"} and a list of the missing variables.
}
\description{
The function checks if specified variables are present in the dataset and presents an error message when specified variables are not present. This function is often used within other CCMHr functions when a specific variable is required to be present.
}
