% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_to_factor.R
\name{sds_to_factor}
\alias{sds_to_factor}
\title{Convert SDS numeric labels to answer option factors}
\usage{
sds_to_factor(data)
}
\arguments{
\item{data}{}
}
\value{
data with all present SDS variables converted to factors
}
\description{
Convert SDS numeric labels to answer option factors
}
\examples{
sds_to_factor(data.frame(SDS_01 = c(3, 2, 4, 1), SDS_02 = c(4, 3, 2, 1)))

}
