% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_to_numeric.r
\name{sds_to_numeric}
\alias{sds_to_numeric}
\title{Convert SDS answer option labels to numeric}
\usage{
sds_to_numeric(data)
}
\arguments{
\item{data}{A CCMH data file with SDS variables}
}
\value{
data with all present SDS variables converted to numeric
}
\description{
Convert SDS answer option labels to numeric
}
