% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first_CCAPS.R
\name{select_first_CCAPS}
\alias{select_first_CCAPS}
\title{Extract the client's first CCAPS administration.}
\usage{
select_first_CCAPS(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = c("UniqueClientID", "CcmhID", "Data_year", "Depression34", "Anxiety34",
    "Social_Anxiety34", "Academics34", "Eating34", "Hostility34", "Alcohol34", "DI",
    "Depression62", "Eating62", "Substance62", "Anxiety62", "Hostility62",
    "Social_Anxiety62", "Family62", "Academics62"),
  by_year = FALSE
)
}
\arguments{
\item{data}{A data frame that contains CCAPS data.}

\item{order_by}{A quoted string to indicate the variable used to determine the order of rows when selecting the first CCAPS administration (row). By default, \code{"Date"}.}

\item{keep_all}{A logical statement to indicate whether all columns should be kept. If \code{TRUE}, all columns are kept. If \code{FALSE}, variable names listed in keep_columns are kept. By default, \code{FALSE}.}

\item{keep_columns}{A quoted string or list of quoted strings of column CCAPS names to retain in the output data frame. By default, \verb{c("UniqueClientID", "CcmhID", "Data_year", "Depression34", "Anxiety34", "Social_Anxiety34", "Academics34", "Eating34", "Hostility34", "Alcohol34", "DI", "Depression62", "Eating62", "Substance62", "Anxiety62", "Hostility62", "Social_Anxiety62", "Family62", "Academics62"}.}

\item{by_year}{A logical argument to indicate if the first CCAPS administration is by year instead of overall. If \verb{TRUE,} the first CCAPS administration for each client will be extracted across each data year (e.g., multiple rows may correspond to a unique client across multiple years). If \code{FALSE}, the first CCAPS administration of each client will be selected regardless of year (e.g., each row should correspond to a single unique client). By default, \code{FALSE}.}
}
\value{
A data frame with the client's first CCAPS administration.
}
\description{
The function extracts the client's first CCAPS administration. It also allows the programmer to pick specific items within CCAPS (e.g., DI, CCAPS_01).
}
