% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first_CCAPS.R
\name{select_first_CLICC}
\alias{select_first_CLICC}
\alias{select_first_clicc}
\title{Extract the client's first CLICC administration.}
\usage{
select_first_CLICC(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = "CLICC",
  by_year = FALSE,
  recode_NA = FALSE
)

select_first_clicc(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = "CLICC",
  by_year = FALSE,
  recode_NA = FALSE
)
}
\arguments{
\item{data}{A data frame that contains CLICC data.}

\item{order_by}{A quoted string to indicate the variable used to determine the order of rows when selecting the first CLICC administration (row). By default, "\code{Date}".}

\item{keep_all}{A logical statement to indicate whether all columns should be kept. If \code{TRUE}, all columns are kept. If \code{FALSE}, variable names listed in keep_columns are kept. By default, \code{FALSE}.}

\item{keep_columns}{A quoted string or list of quoted strings of column CLICC names to retain in the output data frame. By default, \code{"CLICC"} or all CLICC columns.}

\item{by_year}{A logical argument to indicate if the first CLICC administration is by year instead of overall. If \verb{TRUE,} the first CLICC administration for each client will be extracted across each data year (e.g., multiple rows may correspond to a unique client across multiple years). If \code{FALSE}, the first CLICC administration of each client will be selected regardless of year (e.g., each row should correspond to a single unique client). By default, \code{FALSE}.}

\item{recode_NA}{A logical argument to indicate if NA on CLICC should be recoded as 0. If \code{TRUE}, NA will be recoded as 0. By default, \code{FALSE}.}
}
\value{
A data frame with the client's first CLICC administration.
}
\description{
The function extracts the client's first CLICC administration. It also allows the programmer to pick specific items within CLICC (e.g., CLICC_01_01, CLICC_01_02).
}
