% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first_CCAPS.R
\name{select_first_Closure}
\alias{select_first_Closure}
\alias{select_first_closure}
\title{Extract the client's first Case Closure administration.}
\usage{
select_first_Closure(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = "Closure",
  by_year = FALSE,
  recode_NA = FALSE
)

select_first_closure(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = "Closure",
  by_year = FALSE,
  recode_NA = FALSE
)
}
\arguments{
\item{data}{A data frame that contains Case Closure data.}

\item{order_by}{A quoted string to indicate the variable used to determine the order of rows when selecting the first Case Closure administration (row). By default, \code{"Date"}.}

\item{keep_all}{A logical statement to indicate whether all columns should be kept. If \code{TRUE}, all columns are kept. If \code{FALSE}, variable names listed in keep_columns are kept. By default, \code{FALSE}.}

\item{keep_columns}{A quoted string or list of quoted strings of column Case Closure names to retain in the output data frame. By default, \code{"Closure"} or all Case Closure columns.}

\item{by_year}{A logical argument to indicate if the first Case Closure administration is by year instead of overall. If \verb{TRUE,} the first Case Closure administration for each client will be extracted across each data year (e.g., multiple rows may correspond to a unique client across multiple years). If \code{FALSE}, the first Case Closure administration of each client will be selected regardless of year (e.g., each row should correspond to a single unique client). By default, \code{FALSE}.}

\item{recode_NA}{A logical argument to indicate if NA on Case Closure should be recoded as 0. If \code{TRUE}, NA will be recoded as 0. By default, \code{FALSE}.}
}
\value{
A data frame with the client's first Case Closure administration.
}
\description{
The function extracts the client's first Case Closure administration. It also allows the programmer to pick specific items within Case Closure (e.g., CLOSURE_01_101, CLOSURE_01_102).
}
