% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first_CCAPS.R
\name{select_first_SDS}
\alias{select_first_SDS}
\alias{select_first_sds}
\title{Extract the client's first SDS administration.}
\usage{
select_first_SDS(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = "SDS",
  by_year = FALSE
)

select_first_sds(
  data,
  order_by = "Date",
  keep_all = FALSE,
  keep_columns = "SDS",
  by_year = FALSE
)
}
\arguments{
\item{data}{A data frame that contains SDS data.}

\item{order_by}{A quoted string to indicate the variable used to determine the order of rows when selecting the first SDS administration (row). By default, "\code{Date}".}

\item{keep_all}{A logical statement to indicate whether all columns should be kept. If \code{TRUE}, all columns are kept. If \code{FALSE}, variable names listed in keep_columns are kept. By default, \code{FALSE}.}

\item{keep_columns}{A quoted string or list of quoted strings of column SDS names to retain in the output data frame. By default, \code{"SDS"} or all SDS columns.}

\item{by_year}{A logical argument to indicate if the first SDS administration is by year instead of overall. If \verb{TRUE,} the first SDS administration for each client will be extracted across each data year (e.g., multiple rows may correspond to a unique client across multiple years). If \code{FALSE}, the first SDS administration of each client will be selected regardless of year (e.g., each row should correspond to a single unique client). By default, \code{FALSE}.}
}
\value{
A data frame with the client's first SDS administration.
}
\description{
The function extracts the client's first SDS administration. It also allows the programmer to pick specific items within SDS (e.g., SDS_01, SDS_02).
}
