% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_analysis.R
\name{setup_analysis}
\alias{setup_analysis}
\title{Sets up a folder to house a set of analyses.}
\usage{
setup_analysis(
  name,
  data.folder = TRUE,
  data.folder.rawclean = TRUE,
  visuals.folder = TRUE,
  R.folder = TRUE,
  add.date = TRUE,
  add.folder = NULL,
  create.rproj = TRUE
)
}
\arguments{
\item{name}{A quoted string to indicate the name of the main folder. The symbol "/" (forward slashes) is recoded as "_" (underscore).}

\item{data.folder}{A logical argument to add a data subfolder to the main folder. If \code{TRUE}, the subfolder named "data" is added. By default, \code{TRUE}.}

\item{data.folder.rawclean}{A logical argument to add a raw and clean subfolder within the data subfolder. If \code{TRUE}, the subfolders "clean" and "raw" are added. By default, \code{TRUE}.}

\item{visuals.folder}{A logical argument to add a visuals subfolder to the main folder. If \code{TRUE}, the subfolder named "visuals" is added. By default, \code{TRUE}.}

\item{R.folder}{A logical argument to add a R subfolder to the main subfolder. If \code{TRUE}, the subfolder named "R" is added. By default, \code{TRUE}.}

\item{add.date}{A logical argument to add the date to the beginning of the main folder name. Format is Year-Month ("2025_01"). If \code{TRUE}, the date is added to the beginning of the main folder name. By default, \code{TRUE}.}

\item{add.folder}{A quoted string or list of quoted strings with the names of additional folders to be added to the main and subfolders. When creating subfolder(s), make sure the parent folder is listed/created before the subfolder(s). For example, the folder "figures" should be listed before "figures/histograms". By default, \code{NULL}.}

\item{create.rproj}{A logical argument to create a .Rproj file in the main folder. If \code{TRUE}, a R project is added to the folder. By default, \code{TRUE}.}
}
\value{
The function will create a new analysis folder that contains various subfolders. Most subfolder names are standardized (e.g., data, visuals); however, custom subfolders could be created using the add.folder argument.
}
\description{
The function creates folders associated with a specific set of analyses. Primarily, this function is used to create an ad hoc data analysis folder and its corresponding subfolders. To run the function and create a new analysis folder, the programmer must first set their working directory to a specified folder where the analysis folder will be saved. Typically, the working directory is located under the following path: "CCMH - Documents/Data Repository/Ad hoc projects/####", where #### represents the data year. The newly created folder will be saved under this working directory. Most subfolder names are standardized (e.g., data, visuals); however, the programmer may add custom subfolders using the add.folder argument.
}
