% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data_request.R
\name{setup_data_request}
\alias{setup_data_request}
\title{Setup data cleaning file for data request}
\usage{
setup_data_request(requester_last_name)
}
\arguments{
\item{requester_last_name}{The last name of the data requester.}
}
\value{
An R syntax file with the skeleton sections needed to clean data for a data request
}
\description{
Setup data cleaning file for data request
}
\examples{
\dontrun{
setup_data_request(requester_last_name = "Janis")
}
}
