% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_appointment_utilization.r
\name{summarize_appointment_utilization}
\alias{summarize_appointment_utilization}
\title{Summarize appointment utilization}
\usage{
summarize_appointment_utilization(data, by_year = FALSE, recode_NA = FALSE)
}
\arguments{
\item{data}{A data file containing information on individual and group therapy appointments.}

\item{by_year}{A logical statement to indicate if appointment utilization should be summarized by year instead of overall. If \code{TRUE}, the appointment utilization of each client will be summarized by year. If \code{FALSE}, the appointment utilization of each client will be summarized across years. Default = \code{FALSE}.}

\item{recode_NA}{A logical statement to indicate if NA should be recoded to 0. If \code{TRUE}, NA will be recoded to 0. Default = \code{FALSE}.}
}
\value{
A data frame with cleaning appointment utilization data.
}
\description{
Clean and summarize individual and group therapy appointment utilization
}
