% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_clicc.r
\name{summarize_clicc}
\alias{summarize_clicc}
\title{Summarize CLICC data}
\usage{
summarize_clicc(
  data,
  clicc_type = "checkall",
  group_vars = NULL,
  save = FALSE,
  path = "clicc_summary.csv"
)
}
\arguments{
\item{data}{A data file containing clicc.}

\item{clicc_type}{A vector containing the CLICC type. Options include: "checkall" or "topconcern". By default, \code{"checkall"}.}

\item{group_vars}{A vector of grouping variables. By default, \code{NULL}.}

\item{save}{A logical statement indicates whether the table should be saved as a file under a local folder. If false, the table will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway and name if \code{save = TRUE}. By default, \code{"clicc_summary.csv"}.}
}
\value{
Prints demographic tables in concole and/or saves them in under the working directory.
}
\description{
Summarize CLICC data
}
