% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.r
\name{apa_table}
\alias{apa_table}
\alias{APA_table}
\title{Format tables into APA.}
\usage{
apa_table(
  data,
  header.names = NULL,
  title = "Table #",
  subtitle = "Add subtitle",
  save = FALSE,
  path = "table.png",
  width = 1,
  note = NULL
)

APA_table(
  data,
  header.names = NULL,
  title = "Table #",
  subtitle = "Add subtitle",
  save = FALSE,
  path = "table.png",
  width = 1,
  note = NULL
)
}
\arguments{
\item{data}{A data file that contains details of the table.}

\item{header.names}{A list of quoted strings that specifies column or header names. By default, \code{NULL}.}

\item{title}{A quoted string to indicate the title of the table. By default, \code{"Table #"}.}

\item{subtitle}{A quoted string to indicate the subtitle of the table. By default, \code{"Add subtitle"}.}

\item{save}{A logical argument that indicates whether the table should be saved as a file under a local folder. If \code{FALSE}, the table will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway, name, and type if \code{save = TRUE}. By default, \code{"table.png"}.}

\item{width}{A numeric value to indicate the width of the table. By default, \code{1}.}

\item{note}{A quoted string to specify a note at the bottom of the table. By default, \code{NULL}.}
}
\value{
Print and saves APA formatted tables.
}
\description{
This function takes in a table and reformats the table into APA.
}
