% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{A function to create standardized CCMH box plot's}
\usage{
box_plot(
  data,
  x.var,
  y.var,
  group.var = "",
  color = "#838B8B",
  x.order.manual = NULL,
  save = FALSE,
  path = "plot.png",
  plot.width = 20,
  plot.height = 20,
  plot.units = "cm",
  plot.dpi = 320,
  mean.sym = FALSE,
  mean.sym.shape = 3,
  mean.sym.size = 2,
  mean.sym.color = "#000000",
  mean.sym.position = 0.75,
  plot.title = "",
  x.title = "",
  y.title = "",
  plot.title.size = 20,
  axis.title.size = 16,
  axis.label.size = 14,
  text.font = "Avenir",
  y.label.percent = FALSE,
  y.min = NULL,
  y.max = NULL,
  x.wrap = 15,
  x.remove = FALSE,
  caption = FALSE,
  caption.size = 12,
  legend.position = "none",
  legend.order.manual = NULL,
  legend.title = NULL,
  legend.title.size = 16,
  legend.label.size = 14,
  coord.flip = FALSE,
  plot.element1 = NULL,
  plot.element2 = NULL,
  plot.element3 = NULL,
  plot.element4 = NULL,
  plot.element5 = NULL
)
}
\arguments{
\item{data}{A data file containing all variables that will be used in the plot}

\item{x.var}{A quoted string to indicate the variable that will be plotted on the x-axis. The variable class must be a character or factor.}

\item{y.var}{A quoted string to indicate the variable that will be plotted on the y axis. The variable class must be numeric.}

\item{group.var}{A quoted string to indicate the variable that will group the boxes on the x-axis and specify the color of the boxes. The x-axis variable could used to specify colors of each individual boxes. By default, \code{""}.}

\item{color}{A hex code or list of hex codes that indicates the color of the grouped boxes. See Note 1 for more details. By default, \code{"#838B8B"} or a shade of dark grey.}

\item{x.order.manual}{A list of quoted strings to indicate the order of the x axis. By default, \code{NULL}.}

\item{save}{A logical statement indicates whether the plot should be saved as a file under a local folder. If false, the plot will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway and name if \code{save = TRUE}. By default, \code{"plot.png"}.}

\item{plot.width}{A numeric value to indicate the plot's width. By default, \code{30}.}

\item{plot.height}{A numeric value to indicate the plot's height. By default, \code{20}.}

\item{plot.units}{A quoted string to indicate the plot's width and height size units. Options include \verb{in}, \code{cm}, \code{mm}, and \code{px}. By default, \code{cm}.}

\item{plot.dpi}{A numeric value to indicate the plot's image resolution. By default, \code{320}.}

\item{mean.sym}{A logical statement specifies whether a point should be displayed to indicate the mean. By default, \code{FALSE}.}

\item{mean.sym.shape}{A numeric value to indicate the mean point shape. By default, \code{3} or Crosshair.}

\item{mean.sym.size}{A numeric value to indicate the mean point size. By default, \code{2}.}

\item{mean.sym.color}{A hex code to indicate the mean point color. By default, \code{"#000000"} or black.}

\item{mean.sym.position}{A numeric value to indicate the mean point position. By default, \code{0.75}.}

\item{plot.title}{A quoted string to indicate the title of the plot. By default, \code{""}.}

\item{x.title}{A quoted string to indicate the x axis title. By default, \code{""}.}

\item{y.title}{A quoted string to indicate the y axis title. By default, \code{""}.}

\item{plot.title.size}{A numeric value to indicate the plot title text size. By default, \code{20}.}

\item{axis.title.size}{A numeric value to indicate the axis title text size. By default, \code{16}.}

\item{axis.label.size}{A numeric value to indicate the axis labels text size. By default, \code{14}.}

\item{text.font}{A quoted string to indicate the plot's text font. By default, \code{Avenir}.}

\item{y.label.percent}{A logical statement to indicate whether the y axis labels are percentages. By default, \code{FALSE}.}

\item{y.min}{A numeric value to indicate the minimum number presented on the y axis. By default, \code{NULL}.}

\item{y.max}{A numeric value to indicate the maximum number presented on the y axis. By default, \code{NULL}.}

\item{x.wrap}{A numeric value to indicate the number of characters to wrap the x axis text. By default, \code{15}.}

\item{x.remove}{A logical statement to indicate whether x-axis title, labels, and ticks should be removed. By default, \code{FALSE}.}

\item{caption}{A logical statement to indicate whether the CCMH caption should be included in the plot. By default, \code{FALSE}.}

\item{caption.size}{A numeric value to indicate the caption text size. By default, \code{12}.}

\item{legend.position}{A quoted string or numeric vector to indicate the location of the legend. Options include \code{"left"},\code{"top"}, \code{"right"}, \code{"bottom"}, \code{"none"}, or numeric vector c(x,y). By default, \code{"none"}.}

\item{legend.order.manual}{A list of quoted strings to indicate the order of the legend. By default, \code{NULL}.}

\item{legend.title}{A quoted string to indicate the legend title. By default, \code{""}.}

\item{legend.title.size}{A numeric value to indicate the legend title text size. By default, \code{16}.}

\item{legend.label.size}{A numeric value to indicate the legend label text size. By default, \code{14}.}

\item{coord.flip}{A logical statement to indicate if the plot's x and y axis should be flipped. By default, \code{FALSE}.}

\item{plot.element1}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element2}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element3}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element4}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}

\item{plot.element5}{A ggplot plot function and arguments needed for the plot and specified as an object. By default, \code{NULL}.}
}
\value{
A box plot returned as a object or saved under a local file.
}
\description{
Creates standardized CCMH box plot's using ggplot2.
}
\note{
Note 1. The argument \code{color} specifies the color of items within the grouped variable. For example, the group variable items could be the name of automobile companies (i.e., Ford, Dodge, BMV). You can specify color randomly for each company using this list: \code{c("#21501b", "#c51329", "#074e67")}. Or you can specify color directly to each specific company using this list: \code{c("Ford" = "#21501b", "Dodge" = "#c51329", "BMV" = "#074e67")}.

The \code{plot.element} arguments specify additional ggplot2 graphical element(s) needed to complete a specific box plot, but are not specified as an argument in the \code{box_plot} function. For example, the axis label text will always be black unless specified in one of the \code{plot.element} arguments. To change the axis label text color in one of the \code{plot.element} arguments, you would first create an object to represent graphing element (e.g., green.axis.text <- ggplot2::theme(axis.text = ggplot2::element_text(color = "green"))). Then the object would be added to one of the \code{plot.element} arguments (e.g., \code{plot.element01 = green.axis.text}).
}
