% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAPS34_cut_points.R
\name{ccaps34_cuts}
\alias{ccaps34_cuts}
\title{A function to create CCAPS-34 cut points}
\usage{
ccaps34_cuts(data)
}
\arguments{
\item{\code{data}}{A data file containing CCAPS-34 subscales}
}
\value{
A data frame with all the original data in \code{data}, and several additional columns of cut scores coded 1 for above cut score and 0 for below cut score.
}
\description{
Creates binary variables for CCAPS-34 subscales indicating whether the score is above or below the low and high cut respectively, using the 2018 cut points.
}
\note{
If \code{data} does not contain the proper CCAPS subscale names, it will return an error.
}
\examples{
TI1718 <- ccaps34_cuts(TI1718)
}
