% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAPS62_cuts.R
\name{ccaps62_cuts}
\alias{ccaps62_cuts}
\alias{CCAPS62_cuts}
\title{Create CCAPS-62 cut points}
\usage{
ccaps62_cuts(data, version = 2025, first = FALSE)

CCAPS62_cuts(data, version = 2025, first = FALSE)
}
\arguments{
\item{data}{A data file containing CCAPS-62 subscales scores.}

\item{version}{A numeric value to indicate the version of CCAPS cut point scores to use. Options include 2018, 2019, and 2025. By default, \code{2025}.}

\item{first}{A logical argument that indicates if the cut points should be based on variables ending with "\_first" (e.g., Depression62_first; variables created by the CCAPS_change function). If \code{TRUE}, the cut points should be based on variable names ending with "\_first". By default, \code{FALSE}.}
}
\value{
A data frame with all the original data and several additional columns of cut scores.
}
\description{
Creates binary variables for CCAPS-62 subscales indicating whether each subscale score is equal to or below (0) or above (1) the low and high cut scores, respectively.
}
\note{
If the data frame does not contain the proper CCAPS subscale names, it will return an error. Two different variables are created. Variables that end "\_low_cut" (e.g., "Depression62_low_cut") indicate whether or not the client's score was below (0) or above (1) the low cut. Variables that end "\_hi_cut" (e.g., "Depression62_hi_cut") indicate whether or not the client's score was below (0) or above (1) the high cut. The Distress Index (DI) cut scores are not included. To add DI cut scores, a programmer must create a data frame with CCAPS34 cut scores using the ccaps34_cuts function and merge that data frame with a data frame that contains CCAPS62 subscale cut scores.
}
