% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAPS_change.r
\name{CCAPS_change}
\alias{CCAPS_change}
\title{Create a data frame with CCAPS change scores between first and last administration}
\usage{
CCAPS_change(
  data,
  client_identifier = UniqueClientID,
  center_identifier = CcmhID,
  add_items = NA,
  include_first = F,
  include_last = F
)
}
\arguments{
\item{data}{A data frame with CCAPS data.}

\item{client_identifier}{The column uniquely identifying each client. By default, \code{UniqueClientID}.}

\item{center_identifier}{The column uniquely identifying each center. By default \code{CcmhID}.}

\item{add_items}{Specify the columns or CCAPS variables outside of CCAPS subscales to be included in the returned data frame. Default is \code{NA}, resulting in no additional CCAPS items being added to the returned data frame. Multiple items should be listed in the following manner: c("Variable1", "Variable2"). Option \code{"all"} selects all CCAPS 34 items to be in the returned data frame.}

\item{include_first}{A logical argument that specifies if subscale or item scores from the first completion of the CCAPS should be included in the returned data frame. By default \code{FALSE} resulting in CCAPS subscale or specified variable scores of the first administration not being included in the returned data frame. Option \code{"TRUE"} adds first administration scores of CCAPS subscales and specified items to the returned data frame.}

\item{include_last}{A logical argument that specifies if subscale or item scores from the last completion of the CCAPS should be included in the returned data frame. By default \code{FALSE} resulting in CCAPS subscale or specified variable scores of the last administration not being included in the returned data frame. Option \code{"TRUE"} adds last administration scores of CCAPS subscales and specified items to the returned data frame.}
}
\value{
A data frame with CCAPS change scores of subscales and specified CCAPS items. Based on additional arguments, scores from first and last administration will be included as well.
}
\description{
Create a data frame with CCAPS change scores between first and last administration
}
