% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{ccmh_theme}
\alias{ccmh_theme}
\alias{ccmh_caption}
\title{CCMH Plot Theme and Attribution Caption}
\usage{
ccmh_theme(plot_title = 20, axis_title = 16, axis_text = 14)

ccmh_caption()
}
\arguments{
\item{plot_title}{The font size of the plot title}

\item{axis_title}{The font size of the axis title}

\item{axis_text}{The font size of the axis labels}
}
\description{
\code{ccmh_theme} Sets the font to Avenir and adds space between the axis titles and the plot. \cr
\code{ccmh_caption} adds at attribution caption to the plot ("Source: The Center for Collegiate Mental Health")
}
