% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMHr-package.R
\docType{package}
\name{CCMHr-package}
\alias{CCMHr}
\alias{CCMHr-package}
\title{CCMHr: CCMH Data Cleaning}
\description{
Functions used when performing regular data cleaning tasks for CCMH data.
}
\author{
\strong{Maintainer}: Rebecca Janis \email{ruj134@psu.edu}

}
\keyword{internal}
