% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_data_each_year.R
\name{center_data_each_year}
\alias{center_data_each_year}
\title{Selecting only centers who contributed data for all years.}
\usage{
center_data_each_year(data)
}
\arguments{
\item{data}{A data frame with multiple years of CCMH data.}
}
\value{
A data frame with only centers contributing data for all data years included in the original data set.
}
\description{
Selecting only centers who contributed data for all years.
}
\examples{

}
