% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CLICC.R
\name{check_CLICC}
\alias{check_CLICC}
\alias{check_clicc}
\title{Check the response validity of the CLICC items.}
\usage{
check_CLICC(dat)

check_clicc(dat)
}
\arguments{
\item{dat}{A data file containing CLICC items.}
}
\value{
If CLICC_01 and CLICC_03 item responses are valid, the function returns the message, "CLICC looks good!" If CLICC_01 and CLICC_03 item responses are invalid, the function returns a message indicating which CLICC_01 variables had invalid responses and/or the CLICC_03 invalid responses.
}
\description{
Check a data frame to ensure all CLICC_01 items have 1 or NA values or CLICC_03 responses corresponding to possible CLICC numbers (See SDS codebook for possible CLICC_03 values). If these constraints are not satisfied, the function returns a message indicating which CLICC_01 variables had invalid responses and/or the CLICC_03 invalid responses. To run this function, CLICC variable names must be correctly named.
}
