% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colum_classes.R
\name{check_column_classes}
\alias{check_column_classes}
\title{Check column class equality across two data frames.}
\usage{
check_column_classes(dataframe1 = survey, dataframe2 = appts)
}
\arguments{
\item{dataframe1}{A data frame.}

\item{dataframe2}{Another data frame.}
}
\value{
If column classes are equal across two data frames, this message is returned: "No class mismatches detected." If column classes are not equal across two data frames, then a message is returned (i.e., "Class mismatches detected.") along with a printed data frame that details the column, column class for the first data frame, and column class for the second data frame.
}
\description{
A function to check that columns named the same across two data frames have the same classes. To be used before merging two data frames to ensure that the columns with the same names will merge cleanly.
}
