% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clicc_key.R
\docType{data}
\name{clicc_key}
\alias{clicc_key}
\title{CLICC key}
\format{
A data frame with columns:
\describe{
\item{Item}{CLICC item corresponding to the column name of each item in the data}
\item{Concern}{Name of the concern}
\item{Number}{CLICC number corresponding to the top concern in the CLICC_03 item}
}
}
\usage{
clicc_key
}
\description{
Contains a key for the CLICC items and numbers.
}
\keyword{datasets}
