% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CCAPS_bins.R
\name{create_CCAPS_bins}
\alias{create_CCAPS_bins}
\title{Create CCAPS bins for profile curves based on starting CCAPS scores}
\usage{
create_CCAPS_bins(data)
}
\arguments{
\item{data}{A data file with variables for first CCAPS administration named subscale_first}
}
\value{
A data frame with variables for each subscale indicating which bin the client falls in.
}
\description{
Create CCAPS bins for profile curves based on starting CCAPS scores
}
