% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_courses.R
\name{create_courses}
\alias{create_courses}
\title{Create courses of therapy within data}
\usage{
create_courses(data, firstOnly = FALSE)
}
\arguments{
\item{data}{CCMH data files with UniqueClientID and Date variables}

\item{firstOnly}{If FALSE, will return the whole data set. If TRUE, will return only each client's first course.}
}
\value{
A data frame with added variables indicating courses.

If \code{firstOnly} = \code{FALSE}, all courses are retained, with the following variables added:
\itemize{
\item \code{UniqueClientID_byCourse}: an ID unique to each course of therapy. This is created by adding .1 to UniqueClientID for their first course (e.g. 100.1), .2 for their second course (100.2), etc.
\item \code{RankCourse}: The course number that the line of data belongs to, numbered 1 through N of courses.
\item \code{FirstCourse}: A dichotomous variable indicating whether the lines of data belong to a client's first course or not.
}

If \code{firstOnly} = \code{TRUE}, only the first course is selected, and no variables are added:
}
\description{
Create courses of therapy in CCMH data.
}
