% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicate_appointments.R
\name{delete_duplicate_appointments}
\alias{delete_duplicate_appointments}
\title{Delete Duplicate Appointments}
\usage{
delete_duplicate_appointments(
  data,
  client_identifier = "UniqueClientID",
  appointment_identifier = "AppointID"
)
}
\arguments{
\item{data}{A data frame with surveys and appointments.}

\item{client_identifier}{The column uniquely identifying each client. By default, \code{UniqueClientID}.}

\item{appointment_identifier}{The column uniquely identifying each appointment. By default \code{AppointID}, but if this is not available, use \code{Date} instead.}
}
\value{
A data frame with only the first instance of each duplicated appointment.
}
\description{
Delete Duplicate Appointments
}
