% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_tables.r
\name{demographic_tables}
\alias{demographic_tables}
\title{Create demographic tables}
\usage{
demographic_tables(
  data,
  decimal_place = 2,
  text.font = "Avenir",
  print.tables = TRUE,
  save = FALSE,
  path = "demograpics.xlsx"
)
}
\arguments{
\item{data}{A data file containing demographic information.}

\item{decimal_place}{The number of decimal places to round to. Default is \code{2}.}

\item{text.font}{The font type of the text in the table. Default is \code{Avenir}.}

\item{print.tables}{A logical statement indicates whether the table should be printed in the console. By default, \code{TRUE}.}

\item{save}{A logical statement indicates whether the table should be saved as a file under a local folder. If false, the table will be returned as an object. By default, \code{FALSE}.}

\item{path}{A quoted string to indicate the file's pathway and name if \code{save = TRUE}. By default, \code{"table.html"}. File must be an html file.}
}
\value{
Prints demographic tables in concole and/or saves them in under the working directory.
}
\description{
Create CCMH demographic tables.
}
