% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_missing_months.R
\name{detect_missing_months}
\alias{detect_missing_months}
\title{Detects and identifies centers with specific patterns of missing data}
\usage{
detect_missing_months(data)
}
\arguments{
\item{data}{A data file}
}
\value{
A data frame with centers that could have problematic missing data
}
\description{
Detects and identifies centers with missing data for at least one entire month on specified variables
}
\note{
Checks for missing data in variables: Is_appointment, Has_CCAPS, Has_SDS, Has_CLICC, Has_Closure,
and all above variables at during one month
}
