% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_present.R
\name{first_present}
\alias{first_present}
\title{Extract the first value from a vector.}
\usage{
first_present(x, order_by = NULL, default = NULL)
}
\arguments{
\item{x}{A vector.}

\item{order_by}{Another vector to sort \code{x} by.}

\item{default}{A numeric value to use if the \code{x} position does not exist in the input.}
}
\value{
The first non-NA value in a vector.
}
\description{
The function extracts the first non-NA value from a vector.
}
