% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqtab.R
\name{freqtab}
\alias{freqtab}
\title{Frequency table}
\usage{
freqtab(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A frequency table with percentages and cumulative percentages.
}
\description{
Frequency table
}
\examples{
df <- data.frame(y = c(1,1,1,2,2,3))
freqtab(df$y)

}
