% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouply.R
\name{grouply}
\alias{grouply}
\title{A function that combines grouping and a function.}
\usage{
grouply(f, ...)
}
\arguments{
\item{f}{A function.}

\item{...}{A unquoted string to specify a grouping variable.}
}
\value{
A data object.
}
\description{
A function that combines grouping of a data frame by a variable and conducting analysis of the data frame by a function.
}
