% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_present.R
\name{last_present}
\alias{last_present}
\title{Extract the last non-NA value}
\usage{
last_present(x, order_by)
}
\arguments{
\item{x}{A vector}

\item{order_by}{Another vector to sort \code{x} by}
}
\value{
The last non-NA value in \code{x}.
}
\description{
Returns the last non-NA value of \code{x} when sorted by \code{order_by}.
}
\note{
If there are no non-NA values, NA is returned.
}
\examples{
last_present(x = c(NA, 1, 2, NA), order_by = c(1, 2, 3, 4))
last_present(x = c(NA, 1, 2, NA), order_by = c(4, 3, 2, 1))

}
