% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRDa.R
\name{loadRDa}
\alias{loadRDa}
\title{Load an RDA file into a new object name.}
\usage{
loadRDa(fileName)
}
\arguments{
\item{fileName}{The file path to a .rda or .rdata file to be loaded into a new object name.}
}
\value{
A data object.
}
\description{
A function to load .rda and .rdata files with the ability to assign them a new object name instead of the original name.
}
