% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty.R
\name{remove_empty}
\alias{remove_empty}
\title{Exclude columns that have complete missing data.}
\usage{
remove_empty(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame without variables/columns that are completely missing.
}
\description{
Some variables/columns could have complete missing data. This function excludes variables/columns that are completely missing.
}
