% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_free_response.R
\name{remove_free_response}
\alias{remove_free_response}
\title{Remove free response variables}
\usage{
remove_free_response(data)
}
\arguments{
\item{data}{A data object}
}
\value{
The data object without any free response variables
}
\description{
Used to clean data for data requests to ensure that free response data is not shared.
}
