% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_free_response.R
\name{remove_free_response}
\alias{remove_free_response}
\title{Remove free-response CCMH variables.}
\usage{
remove_free_response(data, keep = NULL)
}
\arguments{
\item{data}{A CCMH data frame.}

\item{keep}{A quoted string or list of quoted strings that specify which free response variables should be kept. If \code{NULL}, all free response variables will be removed. By default, \code{NULL}.}
}
\value{
By default, a data frame without the following variables: "SDS_16", "SDS_18", "SDS_21", "SDS_29", "SDS_30", "SDS_35", "SDS_38", "SDS_40", "SDS_43", "SDS_45", "SDS_47", "SDS_49", "SDS_54", "SDS_55", "SDS_89", "SDS_92", "SDS_1047", "SDS_2047", "SDS_3047", "SDS_101", "SDS_129", "Services_Other_Description", "Psychiatric_Services_Other_Description", "ChargeFor_Other_Description", "TherEthnicity_Other", "Highest_Degree_Other", "Highest_Degree_Discipline_Other", "Position_Other", "Grading_Scale_Other_Description", "CCAPSFrequencyOther", "CLOSURE_02", "CLOSURE_05", "CIF_02".
}
\description{
This function removes variables that are free responses.
}
