% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_CCAPS_critical_items.r
\name{rename_CCAPS_critical_items}
\alias{rename_CCAPS_critical_items}
\title{Rename CCAPS critical items}
\usage{
rename_CCAPS_critical_items(data, keep_all = TRUE)
}
\arguments{
\item{data}{A data file.}

\item{keep_all}{A logical statement to indicate if all columns should be keeped. If \code{TRUE} will keep all columns of the orginal data frame and add CCAPS_SI and CCAPS_THO. If S\code{FALSE} will keep columns UniqueClientID, CcmhID, Date, Data_year, CCAPS_SI, and CCAPS_THO critical. Default is \code{TRUE}.}
}
\description{
Rename CCAPS critical items
}
