% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_items.R
\name{required_items}
\alias{required_items}
\title{Displays a error message when specified variables are not present in data}
\usage{
required_items(
  data,
  items,
  message = "The following variable(s) need to be present in data:"
)
}
\arguments{
\item{data}{A data frame.}

\item{items}{Specify the variables that must be present in a data frame to use a function or run an R script. Variables should be listed using the c() function, and each variable name should be quoted. For example, "items = c("Var1", "Var2").}

\item{message}{The message that returns when specified variable(s) are missing. By default, "The following variable(s) need to be present in data:"}
}
\value{
A error message if specified variable is missing.
}
\description{
Displays a error message when specified variables are not present in data
}
