% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_CCAPS.R
\name{score_CCAPS}
\alias{score_CCAPS}
\title{Scores the CCAPS subscales for valid CCAPS administrations.}
\usage{
score_CCAPS(data, CCAPS62 = TRUE)
}
\arguments{
\item{data}{A data file containing CCAPS items.}

\item{CCAPS62}{A logical argument to indicate if the CCAPS62 subscales should be scored. If \code{TRUE}, the CCAPS62 & CCAPS34 subscales are scored. If \code{FALSE}, only the CCAPS34 subscales are scored. By default, \code{TRUE}.}
}
\value{
A data frame with all the original variables and several additional columns, including Has_CCAPS, Is_CCAPS62, Is_ValidCCAPS, Is_ValidCCAPS62, Is_CCAPS34, Is_ValidCCAPS34, Depression34, Anxiety34, Social_Anxiety34, Academics34, Eating34, Hostility34, Alcohol34, DI, Depression62, Eating62, Substance62, Anxiety62, Hostility62, Social_Anxiety62, Family62, and Academics62.
}
\description{
For most CCMH data frames, CCAPS subscales should be scored. However, if this is not the case, this function will score the CCAPS subscales for a valid CCAPS administration. The CCAPS items must be appropriately named (e.g., "CCAPS_01") to use this function. Administrations are considered valid and will be scored if 1/3 or less of the items on any subscale are missing, if less than 1/2 of the data overall is missing, and if the variance of the items is \> 0, indicating that the items were not all marked as the same number. Note that the function does not overwrite CCAPS items to reverse-score them. Instead, it creates a new reverse-scored version, which it then deletes.
}
