% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_CCAPS_screen.R
\name{score_CCAPS_screen}
\alias{score_CCAPS_screen}
\alias{score_ccaps_screen}
\title{Scores the CCAPS subscales for a valid CCAPS screen administration.}
\usage{
score_CCAPS_screen(data)

score_ccaps_screen(data)
}
\arguments{
\item{data}{A data file containing CCAPS screen items.}
}
\value{
A data frame with all the original variables and several additional columns, including Has_CCAPS, Is_ValidCCAPS, Depression, Anxiety, Social_Anxiety, Academics, Eating, Frustration, and Alcohol.
}
\description{
This function scores the CCAPS screen. The CCAPS items must be appropriately named (e.g., "CCAPS_01") to use this function. Administrations are considered valid and will be scored if 1/3 or less of the items on any subscale are missing, if less than 1/2 of the data overall is missing, and if the variance of the items is \> 0, indicating that the items were not all marked as the same number. Note that the function does not overwrite CCAPS items to reverse-score them. Instead, it creates a new reverse-scored version, which it then deletes.
}
