% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_to_factor.R
\name{sds_to_factor}
\alias{sds_to_factor}
\alias{SDS_to_factor}
\title{Convert SDS numeric labels into factors.}
\usage{
sds_to_factor(data)

SDS_to_factor(data)
}
\arguments{
\item{data}{A data file containing SDS items.}
}
\value{
A data frame with all the original variables. Most SDS items will be converted to factors.
}
\description{
When conducting analyses or creating visuals, the SDS items' factor labels are often used instead of the numeric labels to make the results more salient and comprehensible. This function converts the numeric labels of most SDS items into factors.
}
