% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_to_numeric.r
\name{sds_to_numeric}
\alias{sds_to_numeric}
\alias{SDS_to_numeric}
\title{Convert SDS factor labels into numeric.}
\usage{
sds_to_numeric(data)

SDS_to_numeric(data)
}
\arguments{
\item{data}{A data file containing SDS items.}
}
\value{
A data frame with all the original variables. Most SDS items will be converted to numeric.
}
\description{
When conducting analyses, the SDS items' factor labels may need to be converted into a numeric label to conduct the analysis. This function converts the factor labels of most SDS items into numeric values.
}
