% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first.r
\name{select_first}
\alias{select_first}
\title{Extract the client's first administration and merge multiple CCMH forms.}
\usage{
select_first(
  data,
  inner_join_list = NULL,
  left_join_list = NULL,
  order_by = "Date",
  keep_columns_CCAPS = c("UniqueClientID", "CcmhID", "Data_year", "Depression34",
    "Anxiety34", "Social_Anxiety34", "Academics34", "Eating34", "Hostility34",
    "Alcohol34", "DI", "Depression62", "Eating62", "Substance62", "Anxiety62",
    "Hostility62", "Social_Anxiety62", "Family62", "Academics62"),
  keep_columns_SDS = "SDS",
  keep_columns_CLICC = "CLICC",
  keep_columns_Closure = "Closure",
  by_year = FALSE,
  recode_NA_CLICC = FALSE,
  recode_NA_Closure = FALSE
)
}
\arguments{
\item{data}{A data frame that contains CCMH forms.}

\item{inner_join_list}{A quoted string or list of quoted strings to indicate which data forms should be inner joined in the final merged data frame. Options include: "CCAPS", "SDS", "CLICC", "Closure", and "NULL". By default, \code{NULL}.}

\item{left_join_list}{A quoted string or list of quoted strings to indicate which data forms should be left joined in the final merged data frame. Options include: "CCAPS", "SDS", "CLICC", "Closure", and "NULL". By default, \code{NULL}. Data forms listed under inner_join_list can not be listed under left_join_list.}

\item{order_by}{A quoted string to indicate the variable used to determine the order of rows when selecting the first administration (row) across data forms. By default, "Date".}

\item{keep_columns_CCAPS}{A quoted string or list of quoted strings of column CCAPS names to retain in the output data frame. By default, \code{c("UniqueClientID", "CcmhID", "Data_year", "Depression34", "Anxiety34", "Social_Anxiety34", "Academics34", "Eating34", "Hostility34", "Alcohol34", "DI", "Depression62", "Eating62", "Substance62", "Anxiety62", "Hostility62", "Social_Anxiety62", "Family62", "Academics62")}.}

\item{keep_columns_SDS}{A quoted string or list of quoted strings of SDS column names to retain in the output data frame. By default, \code{"SDS"} or all SDS columns.}

\item{keep_columns_CLICC}{A quoted string or list of quoted strings of CLICC column names to retain in the output data frame. By default, \code{"CLICC"} or all CLICC columns.}

\item{keep_columns_Closure}{A quoted string or list of quoted strings of Case Closure column names to retain in the output data frame. By default, \code{"Closure"} or all Case Closure columns.}

\item{by_year}{A logical argument to indicate if the first administration is by year instead of overall. If \verb{TRUE,} the first form administration for each client will be extracted across each data year (e.g., multiple rows may correspond to a unique client across multiple years). If \code{FALSE}, the first administration of each client will be selected regardless of year (e.g., each row should correspond to a single unique client). By default, \code{FALSE}.}

\item{recode_NA_CLICC}{A logical argument to indicate if NA on CLICC should be recoded as 0. If \code{TRUE}, NA will be recoded as 0. By default, \code{FALSE}.}

\item{recode_NA_Closure}{A logical argument to indicate if NA on CLICC should be recoded as 0. If \code{TRUE}, NA will be recoded as 0. By default, \code{FALSE}.}
}
\value{
A data frame with the first administration of multiple forms (based on inner_join_list and left_join_list arguments) and items within forms (based on keep_columns_CCAPS, keep_columns_SDS, keep_columns_CLICC, and keep_columns_Closure arguments).
}
\description{
Many analyses regard a client's first administration on specified forms. For example, analyses may require extracting data from a client's first CCAPS and SDS administration. This function encompasses four processes within a single framework. First, the function allows extracting the first administration on the following forms: CCAPS, SDS, CLICC, and Closure. The extracted forms are determined by the forms listed in the inner_join_list and left_join_list. Second, the function allows the programmer to choose how the forms are merged. Forms listed under inner_join_list are forms required (e.g., if inner_join_list = \code{c("CCAPS", "SDS")}, the client must have both a CCAPS and SDS to be included in the output data frame). Forms listed under left_join_list are forms included but not required (e.g., if \code{left_join_list = "CLICC"}, data on the first administration of the CLICC will be merged, but clients do not need the data on the CLICC to be included in the output data frame). Third, the function allows the programmer to pick specific items within different forms (e.g., the programmer could only extract SDS_01 instead of every SDS item). Lastly, the function allows the recording of NA for the CLICC and Case Closure forms to streamline analyses.
}
