% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_analysis.R
\name{setup_analysis}
\alias{setup_analysis}
\title{Setup data analysis folder}
\usage{
setup_analysis(
  name,
  data.folder = TRUE,
  data.folder.rawclean = TRUE,
  visual.folder = TRUE,
  R.folder = TRUE,
  add.date = TRUE,
  add.folder = NULL,
  add.rfiles = TRUE,
  create.rproj = TRUE
)
}
\arguments{
\item{name}{A string to indicate the name of the main folder.}

\item{data.folder}{A logical statement to add data subfolder to the main folder. Default is \code{TRUE}.}

\item{data.folder.rawclean}{A logical statement to add raw and clean subfolder within the data subfolder. Default is \code{TRUE}.}

\item{R.folder}{A logical statement to add R subfolder to the main subfolder. Default is \code{TRUE}.}

\item{add.date}{A logical statement to add the date to the main folder name. Format is Year-Month (numeric). Default is \code{TRUE}.}

\item{add.folder}{A string or vector list with the names of additional folders to be added to the main and subfolders. When creating a folder that will contain subfolder(s), make sure the main folder is listed before the subfolder(s). Default is \code{NULL}.}

\item{add.rfiles}{A logical statement to add R files based on templates to the R folder. Default is \code{TRUE}.}

\item{create.rproj}{A logical statement to create a .Rproj file in the main folder. Default is \code{TRUE}.}

\item{visuals.folder}{A logical statement to add visuals subfolder to the main folder. Default is \code{TRUE}.}
}
\value{
A analysis folder and R syntax files to conduct data analysis.
}
\description{
Setup data analysis folder
}
