% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data_request.R
\name{setup_data_request}
\alias{setup_data_request}
\title{Setup data cleaning file for data request}
\usage{
setup_data_request(cleaning_folder, requester_name)
}
\arguments{
\item{cleaning_folder}{The folder where the data cleaning file should be saved. Should be formatted "{Requester last name} data cleaning".}

\item{requester_name}{The last name of the data requester.}
}
\value{
An R syntax file with the skeleton sections needed to clean data for a data request
}
\description{
Setup data cleaning file for data request
}
\examples{
\dontrun{
setup_data_request(cleaning_folder = "Janis data cleaning", requester_name = "Janis")
}
}
